Advanced Normalization Tools in R

ANTsR interfaces state of the art image processing with R statistical methods. The project grew out of the need, at University of Pennsylvania, to develop large-scale analytics pipelines that track provenance from scanner to scientific study. ANTsR wraps an ANTs and ITK C++ core via Rcpp to access these frameworks from within R and support reproducible analyses. Specialized functionality in ANTsR includes image segmentation and registration along imaging specific variations of principal component and canonical correlation analysis.

Author

Maintainer: Brian B Avants stnava@gmail.com

Other contributors:

Benjamin M Kandel [contributor]

Jeff T Duda [contributor]

Philip A Cook [contributor]

Nicholas J Tustison [contributor]

Dorian Pustina [contributor]