a basic framework for network analysis that produces graph metrics

An implementation of a network analysis framework for BOLD data. We expect that you mapped a label image ( e.g. aal ) to the 3D BOLD space. We build a network and graph metrics from this image and these labels based on the user-defined graph density level.

antsBOLDNetworkAnalysis(
  bold = NULL,
  mask = NULL,
  labels = NULL,
  motion,
  gdens = 0.2,
  threshLo = 1,
  threshHi = 90,
  freqLo = 0.01,
  freqHi = 0.1,
  winsortrim = 0.02,
  throwaway
)

Arguments

bold

input 4D image

mask

antsImage defines areas of interest

labels

antsImage defines regions of interest ie a parcellation

motion

motion parameters - if missing, will estimate from data

gdens

graph density applied to network covariance matrix

threshLo

lower threshold for the label image

threshHi

upper threshold for the label image

freqLo

lower frequency cutoff

freqHi

upper frequency cutoff

winsortrim

winsorize the bold signal by these values eg 0.02

throwaway

this number of initial bold volumes

Value

list of outputs

Author

BB Avants

Examples

# none yet - this is not very well tested with recent ANTsR
if (FALSE) { # \dontrun{
myimg <- antsImageRead(getANTsRData("ch2"), 3)
mylab <- antsImageRead(getANTsRData("ch2a"), 3)
boldfn <- getANTsRData("pcasl")
bold <- antsImageRead(boldfn, 4)
avgbold <- getAverageOfTimeSeries(bold)
breg <- antsRegistration(avgbold, myimg, typeofTransform = c("AffineFast"))
warpedParcellation <- antsApplyTransforms(avgbold, mylab,
  transformlist = breg$fwdtransforms, interpolator = "NearestNeighbor"
)
mask <- getMask(avgbold)
warpedParcellation <- maskImage(warpedParcellation, img.mask = mask)
old <- NA
labels <- warpedParcellation
gdens <- 0.2
threshLo <- 1
threshHi <- 90
freqLo <- 0.01
freqHi <- 0.1
winsortrim <- 0.02
result <- antsBOLDNetworkAnalysis(bold = bold, mask = mask, warpedParcellation)
} # }