Get Quality Control (QC) Metric Names

This function returns a vector of quality control (QC) metric names used in the antspymm package. These metrics include volume measures, reflection errors, PSNR (Peak Signal-to-Noise Ratio), CNR (Contrast-to-Noise Ratio), and various metrics related to motion correction in rsfMRI and DTI.

antspymm_qc_names()

Value

A character vector containing the names of QC metrics.

Examples

qc_names <- antspymm_qc_names()
print(qc_names)
#>  [1] "T1Hier_resnetGrade"                     
#>  [2] "msk_vol"                                
#>  [3] "T2Flair_msk_vol"                        
#>  [4] "NM1_msk_vol"                            
#>  [5] "NM2_msk_vol"                            
#>  [6] "NM3_msk_vol"                            
#>  [7] "NM4_msk_vol"                            
#>  [8] "NM5_msk_vol"                            
#>  [9] "DTI1_msk_vol"                           
#> [10] "DTI2_msk_vol"                           
#> [11] "rsf1_msk_vol"                           
#> [12] "rsf2_msk_vol"                           
#> [13] "rsf3_msk_vol"                           
#> [14] "reflection_err"                         
#> [15] "T2Flair_reflection_err"                 
#> [16] "NM1_reflection_err"                     
#> [17] "NM2_reflection_err"                     
#> [18] "NM3_reflection_err"                     
#> [19] "NM4_reflection_err"                     
#> [20] "NM5_reflection_err"                     
#> [21] "DTI1_reflection_err"                    
#> [22] "DTI2_reflection_err"                    
#> [23] "rsf1_reflection_err"                    
#> [24] "rsf2_reflection_err"                    
#> [25] "rsf3_reflection_errpsnr"                
#> [26] "T2Flair_psnr"                           
#> [27] "NM1_psnr"                               
#> [28] "NM2_psnr"                               
#> [29] "NM3_psnr"                               
#> [30] "NM4_psnr"                               
#> [31] "NM5_psnr"                               
#> [32] "DTI1_psnr"                              
#> [33] "DTI2_psnr"                              
#> [34] "rsf1_psnr"                              
#> [35] "rsf2_psnr"                              
#> [36] "rsf3_psnr"                              
#> [37] "T1Hier_evratio"                         
#> [38] "T2Flair_wmh_evr"                        
#> [39] "rsfMRI_fcnxpro134_bold_evr"             
#> [40] "rsfMRI_fcnxpro122_bold_evr"             
#> [41] "rsfMRI_fcnxpro129_bold_evr"             
#> [42] "NM2DMT_NM_evr"                          
#> [43] "T2Flair_flair_evr"                      
#> [44] "DTI_dti_fa_evr"                         
#> [45] "cnr"                                    
#> [46] "T2Flair_cnr"                            
#> [47] "NM1_cnr"                                
#> [48] "NM2_cnr"                                
#> [49] "NM3_cnr"                                
#> [50] "NM4_cnr"                                
#> [51] "NM5_cnr"                                
#> [52] "DTI1_cnr"                               
#> [53] "DTI2_cnr"                               
#> [54] "rsf1_cnr"                               
#> [55] "rsf2_cnr"                               
#> [56] "rsf3_cnr"                               
#> [57] "rsfMRI_fcnxpro134_motion_corrected_mean"
#> [58] "rsfMRI_fcnxpro134_high_motion_count"    
#> [59] "rsfMRI_fcnxpro134_high_motion_pct"      
#> [60] "rsfMRI_fcnxpro122_motion_corrected_mean"
#> [61] "rsfMRI_fcnxpro122_high_motion_count"    
#> [62] "rsfMRI_fcnxpro122_high_motion_pct"      
#> [63] "rsfMRI_fcnxpro129_motion_corrected_mean"
#> [64] "rsfMRI_fcnxpro129_high_motion_count"    
#> [65] "rsfMRI_fcnxpro129_high_motion_pct"      
#> [66] "DTI_dti_high_motion_count"              
#> [67] "rsfMRI_fcnxpro134_minutes_original_data"
#> [68] "rsfMRI_fcnxpro134_minutes_censored_data"
#> [69] "rsfMRI_fcnxpro122_minutes_original_data"
#> [70] "rsfMRI_fcnxpro122_minutes_censored_data"
#> [71] "rsfMRI_fcnxpro129_minutes_original_data"
#> [72] "rsfMRI_fcnxpro129_minutes_censored_data"