fitTimeVaryingTransformToPointSets

Estimate a time-varying transform from corresponding point sets (> 2).

fitTimeVaryingTransformToPointSets(
  pointSets,
  timePoints = NULL,
  initialVelocityField = NULL,
  numberOfTimeSteps = NULL,
  domainImage = NULL,
  numberOfFittingLevels = 4,
  meshSize = 1,
  splineOrder = 3,
  displacementWeights = NULL,
  numberOfCompositions = 10,
  compositionStepSize = 0.5,
  numberOfIntegrationSteps = 100,
  sigma = 0,
  convergenceThreshold = 1e-06,
  rasterizePoints = FALSE,
  verbose = FALSE
)

Arguments

pointSets

Corresponding points across sets specified in physical space as a n x d matrix where n is the number of points and d is the dimensionality.

timePoints

Set of scalar values, one for each point-set designating its time position in the velocity flow. If not set, it defaults to equal spacing between 0 and 1.

initialVelocityField

Optional velocity field for initializing optimization. Overrides the number of integration points.

numberOfTimeSteps

Time-varying velocity field parameter. Needs to be equal to or greater than the number of point sets. If not specified, it defaults to the number of point sets.

domainImage

Defines physical domain of the B-spline transform. Must be defined for nonlinear transforms.

numberOfFittingLevels

Integer specifying the number of fitting levels for the B-spline interpolation of the displacement field.

meshSize

Defines the mesh size at the initial fitting level for the B-spline interpolation of the displacement field.

splineOrder

Spline order of the B-spline displacement field.

displacementWeights

vector defining the individual weighting of the corresponding scattered data value. Default = NULL meaning all displacements are weighted the same.

numberOfCompositions

Total number of compositions for the diffeomorphic transform.

compositionStepSize

Scalar multiplication factor of the weighting of the update field for the diffeomorphic transforms.

numberOfIntegrationSteps

Number of steps used for integrating the velocity field.

sigma

Gaussian smoothing standard deviation of the update field (in mm).

convergenceThreshold

Composition-based convergence parameter for the diffeomorphic transforms using a window size of 10 values.

rasterizePoints

Use nearest neighbor rasterization of points for estimating update field (potential speed-up).

verbose

Print progress to the screen.

Value

object containing ANTsR transform, error, and scale (or displacement field)

Author

B Avants, N Tustison