brain segmentation based on geometric priors

uses topological constraints to enhance accuracy of brain segmentation

geoSeg(
  img,
  brainmask,
  priors,
  seginit,
  vesselopt = "none",
  vesselk = 2,
  gradStep = 1.25,
  mrfval = 0.1,
  atroposits = 10,
  jacw = NULL,
  beta = 0.9
)

Arguments

img

input image or list of images (multiple features) where 1st image would typically be the primary constrast

brainmask

binary image

priors

spatial priors, assume first is csf, second is gm, third is wm

seginit

a previously computed segmentation which should have the structure of atropos or kmeansSegmentation output

vesselopt

one of bright, dark or none

vesselk

integer for kmeans vessel-based processing

gradStep

scalar for registration

mrfval

e.g. 0.05 or 0.1

atroposits

e.g. 5 iterations

jacw

precomputed diffeo jacobian

beta

for sigma transformation ( thksig output variable )

Value

list of segmentation result images

Author

Brian B. Avants

Examples

if (FALSE) {
img <- antsImageRead(getANTsRData("simple"), 2)
img <- n3BiasFieldCorrection(img, 4)
img <- n3BiasFieldCorrection(img, 2)
bmk <- getMask(img)
segs <- kmeansSegmentation(img, 3, bmk)
priors <- segs$probabilityimages
seg <- geoSeg(img, bmk, priors)
}