getANTsRData

Downloads antsr test data

getANTsRData(
  fileid,
  usefixedlocation = FALSE,
  verbose = FALSE,
  method = ifelse(Sys.info()["sysname"] == "Linux", "wget", "auto"),
  quiet = FALSE
)

Arguments

fileid

one of the permitted file ids or pass "show" to list all valid possibilities. Note that most require internet access to download.

usefixedlocation

directory to which you download files

verbose

optional boolean

method

Method to be used for downloading files, passed to download.file

quiet

If TRUE, suppress status messages (if any), and the progress bar.

Value

filename string

Author

Avants BB

Examples

fi <- getANTsRData("r16")