Convert an image to the geometric centroids of its signal

Reduces a variate/statistical/network image to a set of centroids describing the center of each stand-alone non-zero component in the image.

getCentroids(img, clustparam = 0)

Arguments

img
the image to reduce to centroids - presumably some kind of statistical or network map

clustparam
look at regions greater than or equal to this size

Value

the centroids are output in matrix of size npoints by 3

Examples

img<-antsImageRead( getANTsRData( "r16" ) )
img<-thresholdImage( img, 90, 120 )
img<-labelClusters( img, 10 )
cents<-getCentroids( img  )