Transform image intensities based on histogram mapping.

Apply B-spline 1-D maps to an input image for intensity warping.

histogramMatchImage2(
  sourceImage,
  referenceImage,
  sourceMask = NULL,
  referenceMask = NULL,
  matchPoints = 64,
  transformDomainSize = 255
)

Arguments

sourceImage

source image.

referenceImage

reference image.

sourceMask

source mask.

referenceMask

reference mask.

matchPoints

parametric points at which the intensity transform displacements are specified between [0, 1]. Alternatively, a single number can be given and the sequence is linearly spaced in [0, 1].

transformDomainSize

Defines the sampling resolution of the B-spline warping.

Value

warped intensity image

Author

Tustison NJ

Examples

library( ANTsR )
sourceImage <- antsImageRead(getANTsRData("r16"), 2)
referenceImage <- antsImageRead(getANTsRData("r64"), 2)
matchedImage <- histogramMatchImage2(sourceImage, referenceImage)