Read Images into a Matrix

Read images into rows of a matrix.

imageListToMatrix(imageList, mask, sigma = NA, epsilon = 0)

Arguments

imageList
A character vector containing a list of image files to read, in order - these are image objects, not file names.

mask
An antsImage containing a binary mask, voxels in the mask are placed in the matrix. If not provided, estimated from first image in list. If the mask is a different size than the image, the images will be downsampled and smoothed to the size of the mask.

sigma
smoothing operation in physical space. See smoothImage.

epsilon
threshold value determining what is included in the mask

Value

A matrix containing the masked data, the result of calling as.numeric(image, mask) on each input image.

See also

Examples

 img <- antsImageRead(getANTsRData('r16') ) %>% resampleImage( c(32,32))
 imglist <- list()
 nvox <- dim(img)[1] * dim(img)[2]
 nsubj <- 5
 for(ii in 1:nsubj){
   imglist[[ ii ]] <- img + rnorm(nvox, sd=mean(img[img!=0]))
 }

#> Constant must have one component or the same number of components as the image
#> Constant must have one component or the same number of components as the image
#> Constant must have one component or the same number of components as the image
#> Constant must have one component or the same number of components as the image
#> Constant must have one component or the same number of components as the image

 mask <- getMask(img)
 imgmat <- imageListToMatrix(imglist, mask)