Multiview PCA

Perform Multiview PCA on multiple datasets with an option for sparse PCA.

multiview_pca(
  views,
  n_components,
  sparse = 0.5,
  max_iter = 100,
  sparsenessAlg = "basic",
  verbose = FALSE
)

Arguments

views

A list of data matrices for each view.

n_components

Number of principal components to compute.

sparse

vector of length views with values between zero and one

max_iter

Maximum number of iterations for the optimization.

sparsenessAlg

NA is default otherwise basic, spmp or orthorank

verbose

Logical, whether to print information about energy and sparsity.

Value

A list containing the common representation Z and transformation matrices W.

Examples

set.seed(123)
n_samples <- 100
n_features_1 <- 50
n_features_2 <- 60
n_features_3 <- 70
n_components <- 5
view1 <- matrix(rnorm(n_samples * n_features_1), nrow = n_samples, ncol = n_features_1)
view2 <- matrix(rnorm(n_samples * n_features_2), nrow = n_samples, ncol = n_features_2)
view3 <- matrix(rnorm(n_samples * n_features_3), nrow = n_samples, ncol = n_features_3)
result <- multiview_pca(list(view1, view2, view3), n_components, sparse = rep(0.5,3), 
 verbose = TRUE)
#> Iteration 1: Energy = 17261.350520
#> Iteration 2: Energy = 16930.496684
#> Iteration 3: Energy = 16581.757324
#> Iteration 4: Energy = 16517.798303
#> Iteration 5: Energy = 16415.096618
#> Iteration 6: Energy = 16385.831513
#> Iteration 7: Energy = 16348.892365
#> Iteration 8: Energy = 16333.879574
#> Iteration 9: Energy = 16318.983585
#> Iteration 10: Energy = 16309.710727
#> Iteration 11: Energy = 16302.467225
#> Iteration 12: Energy = 16295.908445
#> Iteration 13: Energy = 16290.443645
#> Iteration 14: Energy = 16284.856169
#> Iteration 15: Energy = 16279.923018
#> Iteration 16: Energy = 16274.736083
#> Iteration 17: Energy = 16270.529847
#> Iteration 18: Energy = 16266.384021
#> Iteration 19: Energy = 16262.642888
#> Iteration 20: Energy = 16258.766240
#> Iteration 21: Energy = 16256.024385
#> Iteration 22: Energy = 16253.219524
#> Iteration 23: Energy = 16251.086224
#> Iteration 24: Energy = 16248.384725
#> Iteration 25: Energy = 16246.235677
#> Iteration 26: Energy = 16243.969568
#> Iteration 27: Energy = 16242.226217
#> Iteration 28: Energy = 16240.483333
#> Iteration 29: Energy = 16239.184817
#> Iteration 30: Energy = 16237.825838
#> Iteration 31: Energy = 16237.110292
#> Iteration 32: Energy = 16235.893061
#> Iteration 33: Energy = 16234.934730
#> Iteration 34: Energy = 16234.166823
#> Iteration 35: Energy = 16233.456915
#> Iteration 36: Energy = 16232.789672
#> Iteration 37: Energy = 16232.282662
#> Iteration 38: Energy = 16231.788093
#> Iteration 39: Energy = 16230.758136
#> Iteration 40: Energy = 16230.339189
#> Iteration 41: Energy = 16229.798658
#> Iteration 42: Energy = 16229.521586
#> Iteration 43: Energy = 16228.728070
#> Iteration 44: Energy = 16228.627139
#> Iteration 45: Energy = 16228.190209
#> Iteration 46: Energy = 16227.980263
#> Iteration 47: Energy = 16227.767442
#> Iteration 48: Energy = 16227.516399
#> Iteration 49: Energy = 16227.292621
#> Iteration 50: Energy = 16227.034121
#> Iteration 51: Energy = 16226.483355
#> Iteration 52: Energy = 16226.329249
#> Iteration 53: Energy = 16226.121026
#> Iteration 54: Energy = 16225.976208
#> Iteration 55: Energy = 16225.825214
#> Iteration 56: Energy = 16225.702082
#> Iteration 57: Energy = 16225.580423
#> Iteration 58: Energy = 16225.631088
#> Iteration 59: Energy = 16225.462079
#> Iteration 60: Energy = 16225.345351
#> Iteration 61: Energy = 16225.228277
#> Iteration 62: Energy = 16225.103658
#> Iteration 63: Energy = 16225.040506
#> Iteration 64: Energy = 16224.892133
#> Iteration 65: Energy = 16224.780573
#> Iteration 66: Energy = 16224.671463
#> Iteration 67: Energy = 16224.549035
#> Iteration 68: Energy = 16224.431764
#> Iteration 69: Energy = 16224.303827
#> Iteration 70: Energy = 16224.168252
#> Iteration 71: Energy = 16224.032100
#> Iteration 72: Energy = 16223.930799
#> Iteration 73: Energy = 16223.808026
#> Iteration 74: Energy = 16223.703574
#> Iteration 75: Energy = 16223.527278
#> Iteration 76: Energy = 16223.400808
#> Iteration 77: Energy = 16223.265822
#> Iteration 78: Energy = 16223.140526
#> Iteration 79: Energy = 16222.891793
#> Iteration 80: Energy = 16222.842617
#> Iteration 81: Energy = 16222.680448
#> Iteration 82: Energy = 16222.561438
#> Iteration 83: Energy = 16222.410390
#> Iteration 84: Energy = 16222.326099
#> Iteration 85: Energy = 16222.173022
#> Iteration 86: Energy = 16222.037512
#> Iteration 87: Energy = 16221.904782
#> Iteration 88: Energy = 16221.787899
#> Iteration 89: Energy = 16221.665285
#> Iteration 90: Energy = 16221.809375
#> Iteration 91: Energy = 16221.059782
#> Iteration 92: Energy = 16221.097358
#> Iteration 93: Energy = 16220.671956
#> Iteration 94: Energy = 16220.517432
#> Iteration 95: Energy = 16220.264890
#> Iteration 96: Energy = 16220.094766
#> Iteration 97: Energy = 16219.896190
#> Iteration 98: Energy = 16219.730425
#> Iteration 99: Energy = 16219.587645
#> Iteration 100: Energy = 16219.443480
print(result)
#> $Z
#>                [,1]         [,2]         [,3]         [,4]         [,5]
#>   [1,] -0.169588619  0.094499901  0.423575898 -0.217276874 -0.086612273
#>   [2,]  0.142756111 -0.091530461  0.051754313  0.065766720  0.174524771
#>   [3,] -0.007434626 -0.264535195 -0.165726304 -0.112090675  0.230529271
#>   [4,]  0.070644052 -0.011175287 -0.251159388 -0.055031866 -0.040443400
#>   [5,]  0.045241483 -0.082713080  0.375362413  0.023955328  0.058313475
#>   [6,]  0.111908743  0.314642129 -0.231175031 -0.114146725  0.161191208
#>   [7,]  0.156446691  0.066135269  0.299391738  0.087325209 -0.047508613
#>   [8,]  0.031353153  0.136660071  0.038647050  0.158180932 -0.138701592
#>   [9,] -0.016558863  0.257946178  0.112844267  0.065066346  0.146218645
#>  [10,] -0.323359665  0.097766074  0.001016685 -0.124637115 -0.062244865
#>  [11,] -0.106074390  0.195004105 -0.495282039  0.038142404  0.077449823
#>  [12,] -0.021833809  0.273999802 -0.216975175  0.230191873 -0.069422087
#>  [13,]  0.043442461  0.003453013  0.197166241 -0.110935993 -0.155282972
#>  [14,]  0.105751110  0.063734807  0.117284802  0.277078100  0.065761830
#>  [15,] -0.008253494 -0.223506866 -0.157245910  0.029406029  0.101458504
#>  [16,]  0.150018044  0.048698985 -0.160581945 -0.161879820 -0.082710379
#>  [17,]  0.015800101  0.026941473 -0.091620253 -0.080119373  0.104667553
#>  [18,]  0.112823192 -0.480623412  0.335898366  0.314755800 -0.228304621
#>  [19,]  0.137662048  0.055643782 -0.110401334  0.172705915  0.013130391
#>  [20,] -0.152674183  0.040451006 -0.216866727  0.262438924 -0.032947075
#>  [21,]  0.080625605 -0.086815233 -0.199047163  0.163425394  0.220993713
#>  [22,] -0.139742549 -0.028215772  0.128472697 -0.010411133  0.019130064
#>  [23,]  0.076423636  0.256434401  0.028621120  0.135341275  0.068155472
#>  [24,]  0.132918135  0.051402996  0.148617488  0.186321589 -0.031681231
#>  [25,]  0.171021130 -0.193190614  0.397201699 -0.176343689  0.179383402
#>  [26,]  0.319566696 -0.098024241 -0.171718729 -0.428625242  0.203036218
#>  [27,]  0.319695741  0.225914620 -0.010548835 -0.113515788  0.098830898
#>  [28,] -0.299853780 -0.097498677 -0.069014761 -0.138077210  0.160196671
#>  [29,]  0.076475173  0.039278107  0.257545244 -0.383922489  0.121144352
#>  [30,] -0.154633093 -0.183606619  0.026492702 -0.127707726 -0.084110277
#>  [31,]  0.406847053  0.087500782 -0.172569895  0.030743456 -0.101841652
#>  [32,] -0.066772957 -0.152454290 -0.109482053  0.076323520 -0.056290783
#>  [33,] -0.167645038 -0.126242189  0.012419267 -0.053463334 -0.088650840
#>  [34,] -0.005677753 -0.083532445  0.245156726  0.054188361 -0.182398108
#>  [35,]  0.120279204 -0.167394048 -0.144768602 -0.147250133  0.135878471
#>  [36,]  0.068001451 -0.146965374  0.362938881  0.103092318  0.055503253
#>  [37,]  0.088953533 -0.044263784 -0.098239234  0.080191111  0.070469816
#>  [38,] -0.001736313  0.044493278  0.263632961  0.010354872  0.200866273
#>  [39,]  0.166191067 -0.002333946 -0.081021476 -0.010970512 -0.158220726
#>  [40,]  0.151127699  0.257089903 -0.172411043  0.262092506 -0.157471464
#>  [41,] -0.125317494  0.165758860 -0.102694501  0.044467550  0.161214692
#>  [42,]  0.168133082  0.045892914  0.068184506  0.035220608 -0.166351775
#>  [43,] -0.158880550 -0.157415969 -0.018027350 -0.011130555  0.198788820
#>  [44,] -0.014874706  0.340251691 -0.141404633 -0.007768044  0.054704874
#>  [45,] -0.092293143  0.006003580  0.036740811  0.055503431 -0.083878026
#>  [46,] -0.050503155 -0.443735299 -0.017163577 -0.054895733 -0.144356516
#>  [47,] -0.341376552  0.163190130 -0.021655719  0.243528915 -0.175081937
#>  [48,]  0.165358065 -0.143144134  0.007854969 -0.145016874 -0.248692662
#>  [49,]  0.079048660 -0.038661900 -0.375532083 -0.256939086 -0.041158076
#>  [50,] -0.296389539  0.001217429 -0.000786921 -0.274370273  0.190005910
#>  [51,]  0.041896645 -0.215226520 -0.125190412  0.048650798  0.057776890
#>  [52,] -0.065457989 -0.327827787 -0.067600827 -0.032113785 -0.011448996
#>  [53,] -0.084078751  0.446274483  0.133636074 -0.358560095  0.184304961
#>  [54,]  0.241235792  0.220007860 -0.161407443  0.121338906  0.358906051
#>  [55,]  0.044023793  0.036664753 -0.219219376 -0.415681827 -0.318388257
#>  [56,] -0.037695001  0.167185691 -0.032910028  0.405081950  0.110286020
#>  [57,]  0.047655469  0.007186294  0.057934893 -0.051866916 -0.168625533
#>  [58,] -0.124918238  0.132046781  0.026553644  0.019297138  0.143626748
#>  [59,] -0.022566344 -0.138184695 -0.020000933  0.301787482  0.054594152
#>  [60,] -0.006553028  0.169155687 -0.006151207 -0.194243440 -0.214096693
#>  [61,] -0.035657320  0.152963726 -0.135702047  0.220025706  0.042263000
#>  [62,]  0.176752858  0.002988318 -0.104651302 -0.128043092 -0.262309868
#>  [63,] -0.077154759  0.245064461 -0.015378154 -0.068146309 -0.045813418
#>  [64,] -0.115058869  0.009000322 -0.101616517 -0.232986435 -0.241520588
#>  [65,] -0.017097468  0.193019063  0.037752253  0.229240478 -0.163689515
#>  [66,] -0.343138743 -0.200450897 -0.112319466  0.009625197  0.035273528
#>  [67,]  0.012381071  0.278950668  0.296660432 -0.063055594 -0.009184349
#>  [68,]  0.068068392 -0.101592639 -0.054009058  0.124089362 -0.222817011
#>  [69,]  0.038925289  0.200572810 -0.048328196 -0.172567236 -0.231927059
#>  [70,]  0.059528207  0.201410151 -0.130416613 -0.249654536 -0.163500533
#>  [71,]  0.130016334 -0.239115081 -0.034102791 -0.353283709  0.069446497
#>  [72,] -0.465495008 -0.329318057  0.076903234  0.151326004  0.210234585
#>  [73,] -0.191954753  0.310475825 -0.290294995 -0.070073626 -0.214188520
#>  [74,] -0.118728444 -0.073806358 -0.261816903  0.123628211  0.012837530
#>  [75,] -0.139107315 -0.031621517  0.096016307  0.061922952 -0.231853051
#>  [76,]  0.025100635  0.513933308 -0.007033802 -0.020019748 -0.091793506
#>  [77,] -0.075168232 -0.097741623  0.125148401  0.028524504 -0.068779838
#>  [78,]  0.328992751  0.026177592 -0.010314019 -0.025795859 -0.056220488
#>  [79,] -0.074479521  0.203704199  0.226331910  0.197263783 -0.163105817
#>  [80,]  0.026406675 -0.161886425 -0.181367684 -0.035988973 -0.254582316
#>  [81,]  0.048816082 -0.145465553  0.178700613  0.124897273  0.290980240
#>  [82,] -0.011912655  0.017876471 -0.191802368  0.168649480 -0.069199758
#>  [83,] -0.134534924  0.032595850  0.047670452  0.013463902 -0.023988890
#>  [84,] -0.132334022  0.006018499  0.099411565  0.137589739 -0.247694657
#>  [85,] -0.013345233 -0.078830961  0.062848916 -0.041772975  0.128655922
#>  [86,] -0.084421929 -0.091013906 -0.091786836  0.022662767  0.493553235
#>  [87,]  0.139465989 -0.514619469 -0.276523049  0.090402329  0.168754254
#>  [88,] -0.241143823 -0.156570758 -0.201051168 -0.038853209 -0.112171057
#>  [89,]  0.054455106 -0.154915787  0.110696940  0.028039391 -0.251687916
#>  [90,]  0.020598659  0.090255366  0.074130614  0.050579184  0.076586469
#>  [91,] -0.236228416 -0.151633870 -0.165792753  0.108227364 -0.043667256
#>  [92,] -0.038517575 -0.026926065  0.168782159 -0.214975598  0.031416979
#>  [93,]  0.039511619  0.171164360 -0.027489572 -0.119133933  0.100298612
#>  [94,] -0.170613107  0.115137284  0.044955058  0.132306383 -0.079905639
#>  [95,]  0.059243774  0.080610460  0.170395238  0.171521832  0.050310797
#>  [96,] -0.184549583 -0.026852414 -0.096062801 -0.073943311 -0.376172352
#>  [97,]  0.034296069 -0.285344987 -0.032973107 -0.097317494 -0.042716651
#>  [98,]  0.114036314  0.024330372  0.124251117  0.071064580 -0.036119089
#>  [99,]  0.252327360 -0.224974837 -0.011699070  0.235876817 -0.306119321
#> [100,]  0.300545405 -0.174001938 -0.265314314  0.233662433  0.010952408
#> 
#> $W
#> $W[[1]]
#>            [,1]      [,2]       [,3]       [,4]        [,5]
#>  [1,] 0.3115498 1.1045502 -1.0497825 -0.2061484  0.00000000
#>  [2,] 0.0000000 0.0000000 -0.7915643 -0.8251189 -0.52280989
#>  [3,] 0.0000000 0.0000000  0.0000000 -0.8716602 -0.29950485
#>  [4,] 0.0000000 1.4046915 -0.9119748 -0.8218722  0.00000000
#>  [5,] 0.8211825 0.0000000  0.0000000 -0.1227795  0.00000000
#>  [6,] 2.1356606 0.0000000 -0.6527741 -1.7354028  0.00000000
#>  [7,] 0.0000000 0.0000000  0.0000000 -1.6503892 -0.81430618
#>  [8,] 0.1918533 1.4633571 -1.2780161  0.0000000 -0.01243137
#>  [9,] 0.6700165 0.0000000  0.0000000 -0.1046282 -1.76159917
#> [10,] 1.1155502 0.0000000 -0.7169556 -1.5614871  0.00000000
#> [11,] 0.0000000 1.2490035 -2.2566785  0.0000000 -0.08023534
#> [12,] 1.1660827 0.5105930  0.0000000 -1.6183412  0.00000000
#> [13,] 0.0000000 0.0000000 -1.2654995 -1.7361682 -0.02388120
#> [14,] 1.4659757 1.5250952 -0.8237990  0.0000000 -0.61755658
#> [15,] 0.6827833 0.0000000 -0.2186541  0.0000000  0.00000000
#> [16,] 2.0946639 0.5117419  0.0000000 -0.7855303 -0.32043395
#> [17,] 0.0000000 1.1370854  0.0000000  0.0000000  0.00000000
#> [18,] 0.0000000 1.3881477 -0.3376566  0.0000000 -0.53368337
#> [19,] 0.0000000 0.0000000  0.0000000 -0.7117034  0.00000000
#> [20,] 0.0000000 0.0000000 -0.6226299 -2.8077555  0.00000000
#> [21,] 0.0000000 0.6762512  0.0000000  0.0000000  0.00000000
#> [22,] 0.7044144 0.0000000  0.0000000 -0.2252216 -0.23154545
#> [23,] 0.0000000 0.2325514  0.0000000 -0.2641450  0.00000000
#> [24,] 1.9873884 0.8718864  0.0000000  0.0000000 -0.87624779
#> [25,] 0.0000000 0.0000000  0.0000000  0.0000000 -0.17510514
#> [26,] 0.4864972 0.8184087 -0.3547472  0.0000000 -1.06551112
#> [27,] 0.7857280 1.4249471  0.0000000 -1.3118706  0.00000000
#> [28,] 1.3623809 0.4483680 -0.9161522  0.0000000  0.00000000
#> [29,] 0.4318371 0.0000000  0.0000000 -1.1103605 -1.11680985
#> [30,] 0.0000000 0.0000000 -2.1830765  0.0000000 -1.47181801
#> [31,] 0.4742796 0.8532808  0.0000000  0.0000000 -0.34378840
#> [32,] 0.0000000 0.0000000  0.0000000 -0.6886931 -0.23499921
#> [33,] 0.0000000 0.0000000 -0.6658066  0.0000000  0.00000000
#> [34,] 0.3572088 0.0000000 -0.6886199 -0.7102792  0.00000000
#> [35,] 0.0000000 0.0000000 -1.4529670  0.0000000 -0.55108665
#> [36,] 0.0000000 0.2293552  0.0000000  0.0000000 -3.57733196
#> [37,] 0.4677849 0.6930967  0.0000000  0.0000000  0.00000000
#> [38,] 0.5650604 0.0000000  0.0000000 -0.1039092  0.00000000
#> [39,] 0.7353615 0.6222694 -1.1058506 -0.6618284  0.00000000
#> [40,] 0.9957171 1.9968401 -0.3767667 -0.1228071 -0.75942067
#> [41,] 0.0000000 2.2423054 -2.1236288  0.0000000  0.00000000
#> [42,] 0.0000000 0.0000000  0.0000000  0.0000000 -1.07729415
#> [43,] 0.0000000 0.3231024  0.0000000  0.0000000  0.00000000
#> [44,] 1.2744004 2.2140034  0.0000000  0.0000000  0.00000000
#> [45,] 0.0000000 0.2313577  0.0000000  0.0000000 -0.12035470
#> [46,] 1.2711042 0.0000000  0.0000000 -0.7075017  0.00000000
#> [47,] 0.0000000 0.6133567  0.0000000  0.0000000  0.00000000
#> [48,] 0.0000000 0.6641177  0.0000000 -1.5148144 -2.25900882
#> [49,] 0.2684974 0.0000000 -0.7335450  0.0000000 -0.67132380
#> [50,] 1.6068782 0.0000000  0.0000000 -0.6102956 -2.36828292
#> 
#> $W[[2]]
#>             [,1]      [,2]       [,3]        [,4]       [,5]
#>  [1,]  0.0000000 1.2195329 0.00000000 0.046880433 0.00000000
#>  [2,] -0.2777955 0.5960593 0.00000000 0.000000000 0.36977381
#>  [3,] -2.3028921 0.3472439 0.00000000 0.000000000 0.00000000
#>  [4,] -0.2705739 0.3704337 0.00000000 0.000000000 0.00000000
#>  [5,] -1.5994170 0.0000000 0.00000000 0.000000000 0.86639071
#>  [6,] -0.3044545 0.0000000 0.00000000 2.782655995 0.00000000
#>  [7,]  0.0000000 0.2801653 0.00000000 0.000000000 0.00000000
#>  [8,]  0.0000000 0.0000000 1.08613979 0.377992746 0.61371526
#>  [9,]  0.0000000 2.2832913 0.26850541 0.000000000 1.75871990
#> [10,] -0.8472355 0.3953527 0.59316760 2.871996636 0.00000000
#> [11,] -0.1921524 0.0000000 0.00000000 0.705735201 0.38410431
#> [12,] -0.4201845 0.7121002 0.00000000 0.953872114 0.00000000
#> [13,]  0.0000000 0.0000000 0.00000000 0.886072261 0.00000000
#> [14,]  0.0000000 0.0000000 1.79089314 0.568219437 0.26393498
#> [15,]  0.0000000 0.0000000 0.00000000 0.944940427 0.30795748
#> [16,] -0.1157034 0.5887117 0.00000000 0.608665292 1.67646085
#> [17,] -1.0463380 0.4785618 0.00000000 0.811642161 1.37873169
#> [18,] -1.6913092 0.0000000 0.00000000 0.000000000 1.37373165
#> [19,] -1.3751029 0.0000000 0.54547623 0.955154779 0.00000000
#> [20,]  0.0000000 1.0394407 0.00000000 0.436204666 0.60517861
#> [21,]  0.0000000 0.0000000 0.71168073 0.526084549 0.00000000
#> [22,]  0.0000000 0.4616112 0.85070655 0.000000000 0.00000000
#> [23,]  0.0000000 0.0000000 0.00000000 0.000000000 1.05506539
#> [24,] -1.3285508 0.0000000 0.00000000 2.321293298 0.00000000
#> [25,] -0.1296884 1.9868186 0.69195711 0.000000000 1.54380774
#> [26,]  0.0000000 0.3021908 1.91193096 0.000000000 0.00000000
#> [27,]  0.0000000 0.0000000 1.20177528 0.582352887 0.00000000
#> [28,] -0.9390162 0.4103267 0.05781053 0.000000000 1.09850234
#> [29,] -1.1450514 0.0000000 0.00000000 0.000000000 0.03346860
#> [30,] -0.6944437 2.4119847 0.00000000 0.000000000 0.42605289
#> [31,]  0.0000000 0.7175095 1.17463332 2.033966201 0.00000000
#> [32,] -2.1089602 0.6956338 0.00000000 2.067820403 0.00000000
#> [33,] -0.6562054 0.0000000 1.09387817 0.000000000 0.15511508
#> [34,]  0.0000000 0.0000000 0.00000000 0.158396458 0.00000000
#> [35,]  0.0000000 0.0000000 0.32495784 0.374508523 0.13502589
#> [36,]  0.0000000 0.0000000 0.00000000 0.039633373 1.36086770
#> [37,] -1.5485746 0.2123880 0.00000000 0.000000000 1.01995955
#> [38,]  0.0000000 1.2693955 0.00000000 3.532359127 0.04638064
#> [39,] -0.6717365 0.0000000 0.00000000 0.000000000 0.00000000
#> [40,]  0.0000000 0.0000000 0.50881235 1.237935405 0.00000000
#> [41,]  0.0000000 1.3726215 2.51676246 0.000000000 1.50348983
#> [42,]  0.0000000 0.2839036 0.00000000 0.000000000 0.38995535
#> [43,]  0.0000000 0.2432496 0.00000000 0.000000000 2.10063647
#> [44,]  0.0000000 0.0000000 0.96672062 0.059585538 0.25085257
#> [45,] -1.3596943 0.9289146 0.00000000 0.005622587 0.44965112
#> [46,]  0.0000000 0.0000000 0.08734295 0.000000000 0.00000000
#> [47,] -0.5894063 0.6647735 0.60254811 0.085685159 0.00000000
#> [48,] -0.7418423 0.0000000 0.00000000 0.000000000 0.54577295
#> [49,] -2.4451758 0.0000000 1.37113929 0.314766731 0.00000000
#> [50,] -0.2544122 1.6810911 0.19864238 0.000000000 0.05470380
#> [51,] -0.5723024 1.2261198 1.96490823 0.000000000 0.00000000
#> [52,]  0.0000000 1.0242637 1.15966741 0.000000000 0.00000000
#> [53,]  0.0000000 0.0000000 0.00000000 0.410276899 0.00000000
#> [54,] -0.3090810 1.6442554 0.00000000 0.000000000 0.00000000
#> [55,]  0.0000000 1.4574219 0.00000000 0.352156648 0.00000000
#> [56,]  0.0000000 0.0000000 1.01365042 0.000000000 0.00000000
#> [57,]  0.0000000 0.0000000 0.21507756 0.000000000 0.85645992
#> [58,] -1.8969109 0.0000000 1.53954060 2.235399441 0.00000000
#> [59,]  0.0000000 0.0000000 0.00000000 0.000000000 0.00000000
#> [60,] -2.9115674 0.0000000 0.00000000 0.000000000 0.00000000
#> 
#> $W[[3]]
#>             [,1]       [,2]       [,3]       [,4]       [,5]
#>  [1,] 0.68541067 -1.0558585  0.0000000 0.43096235 1.58882642
#>  [2,] 0.00000000  0.0000000  0.0000000 1.88473000 0.00000000
#>  [3,] 0.00000000  0.0000000 -0.4188264 0.04150958 0.96369333
#>  [4,] 0.08126148  0.0000000  0.0000000 0.00000000 0.00000000
#>  [5,] 1.12412405  0.0000000  0.0000000 1.70124276 0.00000000
#>  [6,] 0.16051308  0.0000000 -2.0925588 1.15151192 0.00000000
#>  [7,] 1.32663832  0.0000000  0.0000000 0.00000000 2.06868580
#>  [8,] 0.59117664 -1.5038806 -0.7248360 0.98701923 0.00000000
#>  [9,] 0.15966140  0.0000000 -0.3629020 0.00000000 2.48238293
#> [10,] 0.00000000 -0.1111839 -0.2539727 0.00000000 0.35621405
#> [11,] 1.19705124  0.0000000 -0.3007757 0.00000000 0.58065566
#> [12,] 1.94433686 -0.2637460  0.0000000 1.96886786 0.00000000
#> [13,] 1.54961460 -0.3753980  0.0000000 0.01471907 1.29197616
#> [14,] 0.81269865 -1.2515972  0.0000000 0.20409841 2.77304482
#> [15,] 1.76571029 -0.2869420  0.0000000 1.36212286 0.00000000
#> [16,] 0.39737235  0.0000000  0.0000000 0.00000000 1.43774216
#> [17,] 0.00000000 -0.3451522 -1.1275803 0.00000000 0.00000000
#> [18,] 0.00000000  0.0000000  0.0000000 1.92007472 1.56269210
#> [19,] 0.00000000  0.0000000  0.0000000 0.00000000 0.00000000
#> [20,] 1.26644051  0.0000000  0.0000000 0.00000000 0.09834513
#> [21,] 0.00000000  0.0000000 -2.4842833 0.64868373 0.00000000
#> [22,] 0.00000000  0.0000000  0.0000000 0.00000000 0.00000000
#> [23,] 0.49357513 -0.9829891  0.0000000 0.00000000 0.98544442
#> [24,] 0.00000000  0.0000000 -0.8869743 0.00000000 0.56882731
#> [25,] 0.93311534 -0.9804826 -2.0884511 0.00000000 0.55047576
#> [26,] 0.00000000 -1.8482577 -0.2779267 0.49419330 0.00000000
#> [27,] 0.00000000  0.0000000 -0.4513651 0.00000000 0.55397238
#> [28,] 0.03400182 -1.6269140  0.0000000 0.00000000 0.00000000
#> [29,] 0.00000000 -1.7215030 -2.6941968 0.48036641 0.00000000
#> [30,] 0.21211585 -0.9210156 -1.0298273 0.00000000 0.00000000
#> [31,] 2.34674794 -1.0971564  0.0000000 1.15846048 0.00000000
#> [32,] 0.42063245  0.0000000 -0.4624102 0.94425343 0.00000000
#> [33,] 0.00000000  0.0000000 -1.2332083 0.32974777 0.00000000
#> [34,] 0.00000000 -1.1795172  0.0000000 0.00000000 0.48049754
#> [35,] 1.77963310  0.0000000  0.0000000 1.13388405 0.00000000
#> [36,] 0.00000000 -0.1179547 -1.6209904 0.84164517 0.00000000
#> [37,] 0.00000000  0.0000000  0.0000000 2.09971185 1.47114885
#> [38,] 0.00000000  0.0000000  0.0000000 0.87225883 0.00000000
#> [39,] 0.00000000 -1.3149907  0.0000000 0.00000000 0.00000000
#> [40,] 0.00000000 -1.4207055 -0.1138867 0.00000000 0.00000000
#> [41,] 0.00000000  0.0000000 -0.1372856 0.82478959 0.90115426
#> [42,] 2.27185317 -0.8241741  0.0000000 1.18708681 0.00000000
#> [43,] 0.67814732  0.0000000  0.0000000 0.00000000 0.00000000
#> [44,] 0.79151751  0.0000000 -0.4132126 0.46537911 0.00000000
#> [45,] 0.00000000 -0.8333465  0.0000000 0.00000000 0.00000000
#> [46,] 0.04892817  0.0000000 -0.7924675 0.21747315 2.18074579
#> [47,] 1.15563239 -0.2814499  0.0000000 0.00000000 3.21921027
#> [48,] 2.20376575  0.0000000  0.0000000 0.00000000 0.00000000
#> [49,] 0.14272398 -0.2442088  0.0000000 0.00000000 0.75093492
#> [50,] 0.00000000 -0.1860110  0.0000000 0.00000000 0.47046271
#> [51,] 0.58404681  0.0000000 -2.5395219 0.00000000 0.00000000
#> [52,] 2.60654518  0.0000000 -0.5189190 0.35410948 0.00000000
#> [53,] 0.00000000  0.0000000 -1.1332245 0.96969290 1.55025788
#> [54,] 0.00000000 -0.5646442  0.0000000 1.55711300 0.00000000
#> [55,] 0.00000000 -1.3790919 -0.3084014 0.14654462 0.42523180
#> [56,] 0.00000000  0.0000000  0.0000000 1.68679452 0.00000000
#> [57,] 0.72348949 -0.7185358 -0.7095810 0.00000000 0.32510313
#> [58,] 0.14771266  0.0000000 -1.0717517 0.00000000 0.74846962
#> [59,] 0.00000000 -1.7824016 -0.2447680 0.00000000 2.93923331
#> [60,] 0.00000000  0.0000000  0.0000000 0.00000000 2.15938448
#> [61,] 0.00000000 -1.4827684  0.0000000 0.00000000 0.00000000
#> [62,] 0.00000000  0.0000000 -1.7252784 0.00000000 0.31160321
#> [63,] 0.00000000 -0.1517998 -2.0230513 0.00000000 0.01816130
#> [64,] 3.61628975 -0.9599417  0.0000000 1.42893677 0.94472420
#> [65,] 0.00000000 -1.3850723 -1.7483755 0.56765097 0.40226788
#> [66,] 0.88457479 -0.7009474 -1.2697834 1.48050458 2.18258749
#> [67,] 0.00000000 -0.6115130 -1.5561795 0.42219325 0.00000000
#> [68,] 0.72801184  0.0000000 -0.6905839 0.00000000 0.18013961
#> [69,] 0.00000000  0.0000000 -0.4945782 0.39629343 0.61053270
#> [70,] 0.00000000 -1.0941485 -0.2419491 0.00000000 1.07191630
#> 
#>