Simple pairwiseImageDistanceMatrix function for images

Output contains the NImages x NImages matrix of c('PearsonCorrelation','Mattes') or any Image Metric values available in iMath. Similarity is computed after an affine registration is performed. You can also cluster the images via the dissimilarity measurement, i.e. the negated similarity metric. So, the estimated dissimilarity is returned in the matrix.

pairwiseImageDistanceMatrix(dim, myFileList,
  metrictype = "PearsonCorrelation", nclusters = NA)

Arguments

dim
imageDimension

myFileList
dd<-'MICCAI-2013-SATA-Challenge-Data/CAP/training-images/' myFileList<-list.files(path=dd, pattern = glob2rx('*nii.gz'),full.names = T,recursive = T)

metrictype
similarity function

nclusters
integer controlling max number of clusters to search over

Value

raw dissimilarity matrix is output, symmetrized matrix and clustering (optional) in a list

Examples

# NOT RUN {
  # dsimdata<-pairwiseImageDistanceMatrix( 3, imagefilelist, nclusters = 5 )
# }