Compute SUVR from PET time series images

The standardized uptake value (SUV) ratio (SUVR) is a semi-quantative transformation that is performed on a PET image before intepretation. This function will run rigid motion correction on the PET time series before computing SUVR. The implementation also requires an anatomical image and a segmentation that defines the reference region. This function will produce a mask, a motion corrected image and a PET suvr image mapped to the reference anatomical space.

petSUVR(
  petTime,
  anatomicalImage,
  anatomicalSegmentation,
  smoothingParameter = 2.5,
  labelValue = 0,
  subtractBackground = FALSE,
  mapToPet = TRUE,
  noMotionCorr = FALSE
)

Arguments

petTime

pet time series antsImage or 3D image

anatomicalImage

antsImage that indicates the target anatomical image. one should choose this image carefully, as it will impact results. one might compare results with, for example, either whole head or whole brain targets.

anatomicalSegmentation

antsImage in the same space as anatomicalImage. This image should contain a whole brain mask in addition to a labeled reference region for computing SUVR.

smoothingParameter

physical space smoothing parameter see smoothImage

labelValue

the integer value of the reference region as it appears in anatomicalSegmentation. If not set, we assume a binary segmentation.

subtractBackground

if TRUE, the method will attempt to estimate the background (non-anatomical) activation levels before computing SUVR. In this case, SUVR will be computed after subtracting this background value from the mean activation image.

mapToPet

boolean option causing pet to be used as fixed image in the rigid registration between anatomical and pet

noMotionCorr

boolean option activating simple and fast approach that may be better for dynamic data with insufficient per-frame information

Value

suvr antsImage

Author

Avants BB

Examples

if (FALSE) {
suvr <- petSUVR(petTime, seg)
}