Visualize a regression result by a d3 network visualization.

Use either a force directed graph or a Sankey graph to show relationships between predictors and outcome variables. correlateMyOutcomes should correspond to the outcome variables ...

regressionNetworkViz(mylm, sigthresh = 0.05, whichviz = "Sankey",
  outfile = "temp.html", mygroup = 0, logvals = TRUE, verbose = FALSE,
  correlateMyOutcomes = NA, corthresh = 0.9, zoom = F, doFDR = TRUE)

Arguments

mylm
lm model output from bigLMStats

sigthresh
significance threshold

whichviz
which visualization method

outfile
significance threshold

mygroup
color each entry by group membership

logvals
bool

verbose
bool

correlateMyOutcomes
not sure, see code

corthresh
correlation threshold

zoom
zooming factor

doFDR
bool

Value

html file is output

Examples

# NOT RUN {
colnames(brainpreds)<-paste('Vox',c(1:ncol(brainpreds)),sep='')
colnames( mylm$beta.pval )<-colnames(brainpreds)
demognames<-rownames(mylm$beta.pval)
myout = regressionNetworkViz( mylm , sigthresh=0.05, outfile='temp2.html')
# }