Generate Parameter Grid for SIMLR Search

This function creates and optionally subsets all possible combinations of input parameters for use in a SIMLR grid search.

simlr.parameters(
  nsimlr_options,
  prescaling_options,
  objectiver_options,
  mixer_options,
  sparval_options,
  expBeta_options,
  positivities_options,
  optimus_options,
  constraint_options = list("none"),
  num_samples = 10,
  search_type = c("random", "deterministic", "full")
)

Arguments

nsimlr_options

A list of options for the `nsimlr` parameter.

prescaling_options

A list of options for the `prescaling` parameter.

objectiver_options

A list of options for the `objectiver` parameter.

mixer_options

A list of options for the `mixer` parameter.

sparval_options

A list of options for the `sparval` parameter.

expBeta_options

A list of options for the `ebber` parameter.

positivities_options

A list of options for the `pizzer` parameter.

optimus_options

A list of options for the `optimus` parameter.

constraint_options

A list of options for the `constraint` parameter, default is `list("none")`.

Value

A list containing all (or a subset of) combinations of the provided parameters. Each row in the data frame represents a unique combination of the parameters.

The columns of the returned data frame include:

nsimlr: Values corresponding to the `nsimlr` parameter.

prescaling: Values corresponding to the `prescaling` parameter.

objectiver: Values corresponding to the `objectiver` parameter.

mixer: Values corresponding to the `mixer` parameter.

constraint: Values corresponding to the `constraint` parameter.

sparval: Values corresponding to the `sparval` parameter.

ebber: Values corresponding to the `ebber` parameter.

pizzer: Values corresponding to the `pizzer` parameter.

optimus: Values corresponding to the `optimus` parameter.

Each row represents a specific set of parameters that can be passed to the `simlr.search` function for evaluation. The returned data frame is intended for use in parameter tuning and optimization.