Assess Significance of SiMLR Call

This function performs permutation tests to assess the significance of a SiMLR analysis. For more detail on input parameters, see the original function.

simlr.perm(
  voxmats,
  smoothingMatrices,
  iterations = 10,
  sparsenessQuantiles,
  positivities,
  initialUMatrix,
  mixAlg = c("svd", "ica", "avg", "rrpca-l", "rrpca-s", "pca", "stochastic"),
  orthogonalize = FALSE,
  repeatedMeasures = NA,
  lineSearchRange = c(-1e+10, 1e+10),
  lineSearchTolerance = 1e-08,
  randomSeed,
  constraint = c("none", "Grassmann", "Stiefel"),
  energyType = c("cca", "regression", "normalized", "ucca", "lowRank",
    "lowRankRegression"),
  vmats,
  connectors = NULL,
  optimizationStyle = c("lineSearch", "mixed", "greedy"),
  scale = c("centerAndScale", "sqrtnp", "np", "center", "norm", "none", "impute",
    "eigenvalue", "robust"),
  expBeta = 0,
  jointInitialization = TRUE,
  sparsenessAlg = NA,
  verbose = FALSE,
  nperms = 50,
  FUN = "mean"
)

Arguments

voxmats

A list of voxel matrices.

smoothingMatrices

A list of smoothing matrices.

iterations

Number of iterations. Default is 10.

sparsenessQuantiles

A vector of sparseness quantiles.

positivities

A vector of positivity constraints.

initialUMatrix

Initial U matrix for the algorithm.

mixAlg

The mixing algorithm to use. Default is 'svd'.

orthogonalize

Logical indicating whether to orthogonalize. Default is FALSE.

repeatedMeasures

Repeated measures data. Default is NA.

lineSearchRange

Range for line search. Default is c(-1e+10, 1e+10).

lineSearchTolerance

Tolerance for line search. Default is 1e-08.

randomSeed

Seed for random number generation.

constraint

The constraint type. Default is 'none'.

energyType

The energy type. Default is 'cca'.

vmats

List of V matrices - optional initialization matrices

connectors

List of connectors. Default is NULL.

optimizationStyle

The optimization style. Default is 'lineSearch'.

scale

Scaling method. Default is 'centerAndScale'.

expBeta

Exponential beta value. Default is 0.

jointInitialization

Logical indicating joint initialization. Default is TRUE.

sparsenessAlg

Sparseness algorithm. Default is NA.

verbose

Logical indicating whether to print verbose output. Default is FALSE. values > 1 lead to more verbosity

nperms

Number of permutations for significance testing. Default is 50.

FUN

function for summarizing variance explained

Value

A data frame containing p-values for each permutation.