Symmetric multivariate, penalized image-based linear regression model (symilr) for N modalities

SyMILR minimizes reconstruction error across related modalities. That is, SyMILR will reconstruct each modality matrix from a basis set derived from the other modalities. The basis set can be derived from SVD, ICA or a simple sum of basis representations. This function produces dataset-wide multivariate beta maps for each of the related matrices. The multivariate beta maps are regularized by user input matrices that encode relationships between variables. The idea is overall similar to canonical correlation analysis but generalizes the basis construction and to arbitrary numbers of modalities.

symilr(voxmats, smoothingMatrices, iterations = 10, sparsenessQuantiles,
  positivities, initialUMatrix, mixAlg, orthogonalize = TRUE,
  repeatedMeasures = NA, lineSearchRange = c(0, 10),
  lineSearchTolerance = 0.001, randomSeed, verbose = FALSE)

Arguments

voxmats
A list that contains the named matrices. Note: the optimization will likely perform much more smoothly if the input matrices are each scaled to zero mean unit variance e.g. by the scale function.

smoothingMatrices
list of (sparse) matrices that allow parameter smoothing/regularization. These should be square and same order and size of input matrices.

iterations
number of gradient descent iterations

sparsenessQuantiles
vector of quantiles to control sparseness - higher is sparser

positivities
vector that sets for each matrix if we restrict to positive or negative solution (beta) weights. choices are positive, negative or either as expressed as a string.

initialUMatrix
list of initialization matrix size n by k for each modality. Otherwise, pass a single scalar to control the number of basis functions in which case random initialization occurs. One may also pass a single initialization matrix to be used for all matrices. If this is set to a scalar, or is missing, a random matrix will be used.

mixAlg
'svd', 'ica', 'rrpca-l', 'rrpca-s' or 'avg' denotes the algorithm employed when estimating the mixed modality bases

orthogonalize
boolean to control whether we orthogonalize the solutions explicitly

repeatedMeasures
list of repeated measurement identifiers. this will allow estimates of per identifier intercept.

lineSearchRange
lower and upper limit used in optimize

lineSearchTolerance
tolerance used in optimize, will be multiplied by each matrix norm such that it scales appropriately with input data

randomSeed
controls repeatability of ica-based decomposition

verbose
boolean to control verbosity of output

Value

A list of u, x, y, z etc related matrices.

See also

Examples

set.seed(1500)
nsub = 25
npix = c(100,200,133)
nk = 5
outcome = matrix(rnorm( nsub * nk ),ncol=nk)
outcome1 = matrix(rnorm( nsub * nk ),ncol=nk)
outcome2 = matrix(rnorm( nsub * nk ),ncol=nk)
outcome3 = matrix(rnorm( nsub * nk ),ncol=nk)
view1tx = matrix( rnorm( npix[1]  * nk ), nrow=nk )
view2tx = matrix( rnorm( npix[2]  * nk ), nrow=nk )
view3tx = matrix( rnorm( npix[3]  * nk ), nrow=nk )
mat1 = (outcome %*% t(outcome1) %*% (outcome1)) %*% view1tx
mat2 = (outcome %*% t(outcome2) %*% (outcome2)) %*% view2tx
mat3 = (outcome %*% t(outcome3) %*% (outcome3)) %*% view3tx
result = symilr(list( vox = mat1, vox2 = mat2, vox3 = mat3 ),
   initialUMatrix = nk , verbose=TRUE, iterations=5  )

#> initializing with random matrix:  5 columns

#> [1] 1.391035 2.360680 1.242802
#> [1] "myit = 1 data-term 3250.21306214519 Reg 1.66749161847587e-05 tot 3250.21307882011"
#> e 2825.24125575641 3753.51556067023 3171.88237000894
#> [1] 0.9016994 0.9016994 0.9016994
#> [1] "myit = 2 data-term 2169.76605980208 Reg 1.58416017136556e-05 tot 2169.76607564368"
#> e 1995.42402066607 2326.22457044145 2187.64958829873
#> [1] 1.009438 1.032888 1.025908
#> [1] "myit = 3 data-term 2211.99303141008 Reg 1.60497037259795e-05 tot 2211.99304745979"
#> e 1951.20907134024 2383.47129436596 2301.29872852406
#> [1] 1.0043211 0.9754931 0.9715668
#> [1] "myit = 4 data-term 2166.42748759668 Reg 1.61251696699746e-05 tot 2166.42750372185"
#> e 1882.92666237476 2404.33879127069 2212.01700914459
#> [1] 1.003111 1.004247 1.003298
#> [1] "myit = 5 data-term 2160.83234410867 Reg 1.59474822561153e-05 tot 2160.83236005615"
#> e 1967.88318164086 2300.39771175061 2214.21613893453

p1 = mat1 %*% (result$v[[1]])
p2 = mat2 %*% (result$v[[2]])
p3 = mat3 %*% (result$v[[3]])

# compare to permuted data
s1 = sample( 1:nsub)
s2 = sample( 1:nsub)
resultp = symilr(list( vox = mat1, vox2 = mat2[s1,], vox3 = mat3[s2,] ),
   initialUMatrix = nk , verbose=TRUE, iterations=5  )

#> initializing with random matrix:  5 columns

#> [1] 1.391035 2.360680 1.242802
#> [1] "myit = 1 data-term 3222.35367804401 Reg 1.6033511538354e-05 tot 3222.35369407752"
#> e 2757.31806901683 3780.40038705901 3129.34257805619
#> [1] 1.0659954 0.9216132 0.9041055
#> [1] "myit = 2 data-term 3065.31888989045 Reg 1.65933310918985e-05 tot 3065.31890648378"
#> e 2713.63312365308 3329.90546090226 3152.418085116
#> [1] 0.9988125 0.9976307 0.9976741
#> [1] "myit = 3 data-term 3063.93082180422 Reg 1.65933311377807e-05 tot 3063.93083839755"
#> e 2712.74818519651 3328.55275252066 3150.49152769547
#> [1] 0.9988125 0.9976307 0.9976740
#> [1] "myit = 4 data-term 3063.93082180977 Reg 1.65933311375769e-05 tot 3063.9308384031"
#> e 2712.74818520284 3328.55275252368 3150.49152770279
#> [1] 0.9988125 0.9976307 0.9976740
#> [1] "myit = 5 data-term 3063.93082180977 Reg 1.65933311375769e-05 tot 3063.9308384031"
#> e 2712.74818520284 3328.55275252369 3150.49152770279

p1p = mat1 %*% (resultp$v[[1]])
p2p = mat2[s1,] %*% (resultp$v[[2]])
p3p = mat3[s2,] %*% (resultp$v[[3]])

# compare to SVD
svd1 = svd( mat1, nu=nk, nv=0 )$u
svd2 = svd( mat2, nu=nk, nv=0 )$u
svd3 = svd( mat3, nu=nk, nv=0 )$u

# real
range(cor(p1,p2))

#> [1] -0.7955653  0.8157355

range(cor(p1,p3))

#> [1] -0.9444966  0.8492672

range(cor(p3,p2))

#> [1] -0.7307740  0.6799381

# permuted
range(cor(p1p,p2p))

#> [1] -0.4417842  0.4479953

range(cor(p1p,p3p))

#> [1] -0.5256528  0.4614072

range(cor(p3p,p2p))

#> [1] -0.4988755  0.4728194

# svd
print( range(cor( svd1,svd2) ))

#> [1] -0.6848291  0.6203092