Whitening Matrix

This function performs matrix whitening on the input matrix `X` using Singular Value Decomposition (SVD). Whitening transforms the input matrix into one where the covariance matrix is the identity matrix.

whiten_matrix(X)

Arguments

X

A numeric matrix to be whitened. The matrix should have observations as rows and features as columns.

Value

A list containing:

whitened_matrix

The whitened matrix where the covariance matrix is the identity matrix.

whitening_matrix

The whitening transformation matrix used to whiten the input matrix.

Examples

set.seed(123)
X <- matrix(rnorm(1000), nrow = 20, ncol = 50)  # Example with p = 50 and n = 20
result <- whiten_matrix(X)
X_whitened <- result$whitened_matrix
whitening_matrix <- result$whitening_matrix
# Verify that the covariance matrix of the whitened matrix is close to identity
cov_X_whitened <- cov(X_whitened)
print(round(cov_X_whitened, 2))
#>        [,1]  [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10] [,11] [,12]
#>  [1,]  0.13 -0.03 -0.02  0.01 -0.03  0.01  0.02 -0.02 -0.01  0.04  0.02  0.02
#>  [2,] -0.03  0.11  0.02  0.03  0.04  0.02  0.00 -0.01 -0.01  0.02 -0.01 -0.04
#>  [3,] -0.02  0.02  0.12 -0.01 -0.03  0.00  0.04 -0.02  0.05  0.00  0.02 -0.01
#>  [4,]  0.01  0.03 -0.01  0.13 -0.01 -0.02  0.00  0.01  0.00  0.03  0.03 -0.03
#>  [5,] -0.03  0.04 -0.03 -0.01  0.11  0.03 -0.01  0.01  0.01  0.00 -0.01  0.01
#>  [6,]  0.01  0.02  0.00 -0.02  0.03  0.08 -0.04 -0.02  0.00  0.01  0.02 -0.01
#>  [7,]  0.02  0.00  0.04  0.00 -0.01 -0.04  0.15  0.00  0.00  0.00  0.00  0.02
#>  [8,] -0.02 -0.01 -0.02  0.01  0.01 -0.02  0.00  0.15 -0.01  0.01  0.04  0.03
#>  [9,] -0.01 -0.01  0.05  0.00  0.01  0.00  0.00 -0.01  0.15 -0.03  0.02  0.00
#> [10,]  0.04  0.02  0.00  0.03  0.00  0.01  0.00  0.01 -0.03  0.12  0.04  0.00
#> [11,]  0.02 -0.01  0.02  0.03 -0.01  0.02  0.00  0.04  0.02  0.04  0.15  0.01
#> [12,]  0.02 -0.04 -0.01 -0.03  0.01 -0.01  0.02  0.03  0.00  0.00  0.01  0.09
#> [13,]  0.04  0.00 -0.05  0.00  0.04  0.04 -0.03 -0.02 -0.03 -0.01 -0.03  0.01
#> [14,]  0.04 -0.03  0.01 -0.04  0.03  0.01  0.03 -0.04  0.01 -0.02  0.00  0.02
#> [15,]  0.02  0.04 -0.03 -0.01  0.03  0.01 -0.05  0.01 -0.06  0.03 -0.03 -0.01
#> [16,]  0.01 -0.01  0.01  0.03 -0.02  0.03  0.02  0.06 -0.02  0.00  0.00  0.00
#> [17,] -0.03 -0.01  0.02  0.01 -0.01 -0.01 -0.01  0.02  0.06 -0.01 -0.01  0.02
#> [18,] -0.01  0.00  0.00  0.00 -0.03 -0.01 -0.04 -0.04 -0.04  0.00  0.00  0.01
#> [19,]  0.01  0.00  0.04  0.00  0.01  0.00  0.03 -0.01  0.03  0.01  0.01  0.05
#> [20,]  0.02  0.00  0.00  0.05  0.00 -0.01  0.01 -0.04 -0.03  0.03  0.02 -0.01
#> [21,] -0.05  0.03  0.00 -0.01  0.03 -0.03  0.04  0.02  0.02 -0.04 -0.03  0.02
#> [22,] -0.05 -0.03 -0.02 -0.02  0.03 -0.02 -0.05  0.01  0.03 -0.01  0.02  0.01
#> [23,]  0.02  0.01 -0.06 -0.07  0.04 -0.01 -0.02  0.00 -0.01 -0.01 -0.06  0.03
#> [24,]  0.03 -0.02 -0.06  0.03  0.03  0.00  0.00  0.03  0.00  0.01 -0.01  0.01
#> [25,]  0.01 -0.05 -0.03 -0.01 -0.01 -0.02  0.01  0.01 -0.01  0.01  0.00  0.03
#> [26,]  0.02  0.00  0.01  0.00  0.00  0.02 -0.03  0.00  0.03 -0.03  0.03  0.04
#> [27,] -0.01  0.00 -0.01 -0.02  0.03  0.02 -0.02  0.00 -0.01  0.01 -0.02 -0.02
#> [28,]  0.00 -0.01 -0.02  0.02 -0.01 -0.02 -0.03  0.05 -0.03  0.00  0.01  0.03
#> [29,]  0.00  0.02  0.03  0.01  0.03  0.01 -0.03 -0.02  0.03  0.04  0.02  0.00
#> [30,]  0.00 -0.01  0.00  0.01 -0.04 -0.01 -0.06  0.02  0.04 -0.01 -0.02 -0.04
#> [31,] -0.02  0.00 -0.03  0.01 -0.02 -0.02 -0.04  0.00 -0.02 -0.05  0.00 -0.01
#> [32,] -0.01  0.02 -0.03  0.04  0.03 -0.01  0.00 -0.06  0.03 -0.01 -0.05  0.02
#> [33,] -0.02 -0.01 -0.02  0.01  0.03  0.01 -0.04  0.01  0.00  0.00 -0.01  0.01
#> [34,]  0.01 -0.02 -0.06  0.04  0.00  0.02 -0.05 -0.01  0.00 -0.02  0.04  0.02
#> [35,]  0.03  0.00  0.00 -0.01 -0.01 -0.02  0.05  0.06 -0.03  0.07  0.07  0.01
#> [36,]  0.02  0.04  0.01  0.04  0.03  0.02 -0.02  0.03  0.01 -0.02 -0.02  0.00
#> [37,]  0.02  0.02 -0.01 -0.01 -0.01  0.02  0.01 -0.02 -0.03 -0.01 -0.07 -0.02
#> [38,] -0.01  0.03  0.00  0.01  0.02  0.02 -0.03 -0.01  0.00  0.05  0.04  0.02
#> [39,]  0.02  0.01 -0.04  0.06  0.00  0.01  0.01 -0.03 -0.06  0.04  0.01 -0.05
#> [40,]  0.03  0.04 -0.03  0.02  0.00  0.00  0.00  0.01 -0.01  0.01  0.01 -0.01
#> [41,] -0.03 -0.02  0.02  0.02 -0.01 -0.03 -0.02 -0.04  0.01 -0.01  0.00 -0.02
#> [42,]  0.04  0.02  0.05 -0.02 -0.04  0.00  0.01 -0.03  0.02  0.00  0.00  0.02
#> [43,] -0.05  0.01  0.02  0.00 -0.01  0.00 -0.03  0.01 -0.06  0.00  0.00  0.01
#> [44,] -0.02  0.00 -0.01  0.01  0.02  0.03 -0.02  0.02 -0.01 -0.02 -0.01 -0.04
#> [45,]  0.01 -0.02  0.02 -0.03 -0.02  0.03 -0.02 -0.03  0.06 -0.05 -0.01 -0.02
#> [46,]  0.05  0.01  0.02  0.03  0.01 -0.02  0.01  0.00 -0.01  0.02 -0.02  0.02
#> [47,]  0.00 -0.04  0.01  0.04 -0.04 -0.02 -0.03  0.00  0.01  0.01  0.00  0.01
#> [48,] -0.05  0.00 -0.02  0.00  0.01 -0.04 -0.04  0.03  0.01 -0.03 -0.01 -0.02
#> [49,] -0.02 -0.05 -0.01 -0.02  0.00  0.02 -0.03  0.00 -0.03 -0.02  0.02  0.04
#> [50,]  0.00  0.03 -0.03  0.05  0.01 -0.01  0.02 -0.01  0.01 -0.02  0.03 -0.04
#>       [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24]
#>  [1,]  0.04  0.04  0.02  0.01 -0.03 -0.01  0.01  0.02 -0.05 -0.05  0.02  0.03
#>  [2,]  0.00 -0.03  0.04 -0.01 -0.01  0.00  0.00  0.00  0.03 -0.03  0.01 -0.02
#>  [3,] -0.05  0.01 -0.03  0.01  0.02  0.00  0.04  0.00  0.00 -0.02 -0.06 -0.06
#>  [4,]  0.00 -0.04 -0.01  0.03  0.01  0.00  0.00  0.05 -0.01 -0.02 -0.07  0.03
#>  [5,]  0.04  0.03  0.03 -0.02 -0.01 -0.03  0.01  0.00  0.03  0.03  0.04  0.03
#>  [6,]  0.04  0.01  0.01  0.03 -0.01 -0.01  0.00 -0.01 -0.03 -0.02 -0.01  0.00
#>  [7,] -0.03  0.03 -0.05  0.02 -0.01 -0.04  0.03  0.01  0.04 -0.05 -0.02  0.00
#>  [8,] -0.02 -0.04  0.01  0.06  0.02 -0.04 -0.01 -0.04  0.02  0.01  0.00  0.03
#>  [9,] -0.03  0.01 -0.06 -0.02  0.06 -0.04  0.03 -0.03  0.02  0.03 -0.01  0.00
#> [10,] -0.01 -0.02  0.03  0.00 -0.01  0.00  0.01  0.03 -0.04 -0.01 -0.01  0.01
#> [11,] -0.03  0.00 -0.03  0.00 -0.01  0.00  0.01  0.02 -0.03  0.02 -0.06 -0.01
#> [12,]  0.01  0.02 -0.01  0.00  0.02  0.01  0.05 -0.01  0.02  0.01  0.03  0.01
#> [13,]  0.12  0.03  0.00 -0.02 -0.02 -0.03  0.02  0.04 -0.03 -0.03  0.02  0.01
#> [14,]  0.03  0.14  0.01  0.00 -0.03 -0.04  0.02  0.02  0.01  0.00  0.01 -0.03
#> [15,]  0.00  0.01  0.13  0.02 -0.03  0.00 -0.03 -0.03  0.01 -0.02  0.07  0.01
#> [16,] -0.02  0.00  0.02  0.16  0.02 -0.02 -0.02 -0.04  0.00 -0.07 -0.05  0.04
#> [17,] -0.02 -0.03 -0.03  0.02  0.13 -0.03  0.07 -0.02  0.01  0.00  0.02  0.02
#> [18,] -0.03 -0.04  0.00 -0.02 -0.03  0.19 -0.01  0.03 -0.02  0.04 -0.01  0.01
#> [19,]  0.02  0.02 -0.03 -0.02  0.07 -0.01  0.13  0.01  0.01 -0.01  0.01  0.00
#> [20,]  0.04  0.02 -0.03 -0.04 -0.02  0.03  0.01  0.12 -0.04  0.01 -0.05 -0.02
#> [21,] -0.03  0.01  0.01  0.00  0.01 -0.02  0.01 -0.04  0.10 -0.01  0.04  0.00
#> [22,] -0.03  0.00 -0.02 -0.07  0.00  0.04 -0.01  0.01 -0.01  0.17 -0.01  0.01
#> [23,]  0.02  0.01  0.07 -0.05  0.02 -0.01  0.01 -0.05  0.04 -0.01  0.15  0.03
#> [24,]  0.01 -0.03  0.01  0.04  0.02  0.01  0.00 -0.02  0.00  0.01  0.03  0.11
#> [25,]  0.00  0.04 -0.01 -0.01 -0.01 -0.02 -0.02  0.00  0.01  0.05  0.01 -0.02
#> [26,]  0.03  0.02 -0.01 -0.01 -0.03  0.02  0.00  0.01  0.00  0.02  0.00 -0.02
#> [27,]  0.03  0.01 -0.01 -0.02  0.01 -0.01  0.04  0.01 -0.02  0.04 -0.01  0.02
#> [28,]  0.05  0.00  0.03  0.01  0.03 -0.04  0.05  0.01  0.02 -0.03  0.02 -0.01
#> [29,] -0.02 -0.02  0.02 -0.01  0.02  0.04  0.02  0.02 -0.04  0.02  0.00  0.04
#> [30,] -0.03 -0.01  0.02 -0.01  0.02 -0.04 -0.06 -0.04 -0.01  0.03  0.00 -0.05
#> [31,]  0.01  0.03  0.01 -0.04  0.01  0.04  0.00  0.03  0.02  0.03  0.02 -0.04
#> [32,]  0.04  0.00 -0.02 -0.01  0.03  0.01  0.01  0.04  0.03 -0.04  0.02  0.01
#> [33,]  0.03  0.03  0.02  0.01 -0.02 -0.03 -0.04  0.00  0.00  0.00 -0.02 -0.02
#> [34,]  0.01 -0.03  0.00 -0.01  0.03  0.00  0.01 -0.01 -0.01  0.02  0.02  0.03
#> [35,] -0.02 -0.03  0.00 -0.02 -0.03 -0.03 -0.02  0.01 -0.03 -0.01  0.00 -0.01
#> [36,]  0.03  0.01  0.02  0.04 -0.01 -0.02  0.01  0.01 -0.02  0.00 -0.01  0.02
#> [37,]  0.02 -0.02  0.01  0.06 -0.01  0.03 -0.01 -0.02 -0.01 -0.03  0.02  0.05
#> [38,]  0.00 -0.04 -0.01 -0.04 -0.01  0.07  0.05 -0.01  0.01  0.08 -0.01  0.00
#> [39,]  0.03  0.00 -0.01  0.00 -0.04 -0.02 -0.03  0.06 -0.04 -0.01 -0.05  0.00
#> [40,]  0.01 -0.02 -0.01 -0.03 -0.01  0.05  0.00  0.03 -0.02  0.00  0.02  0.00
#> [41,]  0.02 -0.04 -0.02 -0.07 -0.02  0.00  0.00  0.02 -0.01  0.01 -0.03 -0.01
#> [42,] -0.01  0.01  0.00 -0.03  0.03  0.03  0.05  0.01 -0.02 -0.04  0.03 -0.05
#> [43,] -0.04  0.01  0.06  0.02  0.05  0.00  0.01  0.00  0.00  0.00  0.00 -0.04
#> [44,]  0.04 -0.01 -0.02  0.05  0.00 -0.04 -0.04  0.03 -0.03 -0.06 -0.04  0.02
#> [45,]  0.01  0.01 -0.02  0.01  0.01 -0.01 -0.01 -0.04  0.00 -0.03  0.00 -0.01
#> [46,] -0.02  0.05  0.06  0.04 -0.03  0.05  0.02 -0.01  0.00  0.00  0.00  0.03
#> [47,] -0.01  0.01  0.00  0.04  0.05  0.01  0.02  0.03 -0.02  0.01 -0.04  0.00
#> [48,] -0.03 -0.03  0.02 -0.02  0.04  0.02 -0.02  0.00  0.02  0.05  0.03  0.03
#> [49,]  0.02  0.02 -0.03  0.01  0.04  0.06  0.05  0.01 -0.03  0.00 -0.02 -0.01
#> [50,] -0.01  0.01  0.01  0.01 -0.02  0.03  0.00  0.00  0.04 -0.02 -0.01  0.03
#>       [,25] [,26] [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36]
#>  [1,]  0.01  0.02 -0.01  0.00  0.00  0.00 -0.02 -0.01 -0.02  0.01  0.03  0.02
#>  [2,] -0.05  0.00  0.00 -0.01  0.02 -0.01  0.00  0.02 -0.01 -0.02  0.00  0.04
#>  [3,] -0.03  0.01 -0.01 -0.02  0.03  0.00 -0.03 -0.03 -0.02 -0.06  0.00  0.01
#>  [4,] -0.01  0.00 -0.02  0.02  0.01  0.01  0.01  0.04  0.01  0.04 -0.01  0.04
#>  [5,] -0.01  0.00  0.03 -0.01  0.03 -0.04 -0.02  0.03  0.03  0.00 -0.01  0.03
#>  [6,] -0.02  0.02  0.02 -0.02  0.01 -0.01 -0.02 -0.01  0.01  0.02 -0.02  0.02
#>  [7,]  0.01 -0.03 -0.02 -0.03 -0.03 -0.06 -0.04  0.00 -0.04 -0.05  0.05 -0.02
#>  [8,]  0.01  0.00  0.00  0.05 -0.02  0.02  0.00 -0.06  0.01 -0.01  0.06  0.03
#>  [9,] -0.01  0.03 -0.01 -0.03  0.03  0.04 -0.02  0.03  0.00  0.00 -0.03  0.01
#> [10,]  0.01 -0.03  0.01  0.00  0.04 -0.01 -0.05 -0.01  0.00 -0.02  0.07 -0.02
#> [11,]  0.00  0.03 -0.02  0.01  0.02 -0.02  0.00 -0.05 -0.01  0.04  0.07 -0.02
#> [12,]  0.03  0.04 -0.02  0.03  0.00 -0.04 -0.01  0.02  0.01  0.02  0.01  0.00
#> [13,]  0.00  0.03  0.03  0.05 -0.02 -0.03  0.01  0.04  0.03  0.01 -0.02  0.03
#> [14,]  0.04  0.02  0.01  0.00 -0.02 -0.01  0.03  0.00  0.03 -0.03 -0.03  0.01
#> [15,] -0.01 -0.01 -0.01  0.03  0.02  0.02  0.01 -0.02  0.02  0.00  0.00  0.02
#> [16,] -0.01 -0.01 -0.02  0.01 -0.01 -0.01 -0.04 -0.01  0.01 -0.01 -0.02  0.04
#> [17,] -0.01 -0.03  0.01  0.03  0.02  0.02  0.01  0.03 -0.02  0.03 -0.03 -0.01
#> [18,] -0.02  0.02 -0.01 -0.04  0.04 -0.04  0.04  0.01 -0.03  0.00 -0.03 -0.02
#> [19,] -0.02  0.00  0.04  0.05  0.02 -0.06  0.00  0.01 -0.04  0.01 -0.02  0.01
#> [20,]  0.00  0.01  0.01  0.01  0.02 -0.04  0.03  0.04  0.00 -0.01  0.01  0.01
#> [21,]  0.01  0.00 -0.02  0.02 -0.04 -0.01  0.02  0.03  0.00 -0.01 -0.03 -0.02
#> [22,]  0.05  0.02  0.04 -0.03  0.02  0.03  0.03 -0.04  0.00  0.02 -0.01  0.00
#> [23,]  0.01  0.00 -0.01  0.02  0.00  0.00  0.02  0.02 -0.02  0.02  0.00 -0.01
#> [24,] -0.02 -0.02  0.02 -0.01  0.04 -0.05 -0.04  0.01 -0.02  0.03 -0.01  0.02
#> [25,]  0.09  0.02 -0.03  0.02 -0.05  0.03  0.03  0.02  0.03  0.02  0.00 -0.03
#> [26,]  0.02  0.10 -0.05  0.01  0.00 -0.01  0.02  0.03  0.02  0.03 -0.01  0.04
#> [27,] -0.03 -0.05  0.11  0.01  0.02 -0.01  0.01 -0.06 -0.03 -0.05 -0.02  0.02
#> [28,]  0.02  0.01  0.01  0.15 -0.04 -0.01  0.06 -0.02  0.01  0.01 -0.02  0.01
#> [29,] -0.05  0.00  0.02 -0.04  0.11 -0.03 -0.05  0.01 -0.01 -0.02  0.01  0.03
#> [30,]  0.03 -0.01 -0.01 -0.01 -0.03  0.17  0.04 -0.02  0.05  0.01 -0.01  0.01
#> [31,]  0.03  0.02  0.01  0.06 -0.05  0.04  0.15  0.00 -0.01  0.01 -0.04 -0.01
#> [32,]  0.02  0.03 -0.06 -0.02  0.01 -0.02  0.00  0.17  0.05  0.03 -0.05 -0.01
#> [33,]  0.03  0.02 -0.03  0.01 -0.01  0.05 -0.01  0.05  0.10  0.01 -0.02  0.00
#> [34,]  0.02  0.03 -0.05  0.01 -0.02  0.01  0.01  0.03  0.01  0.17 -0.02 -0.01
#> [35,]  0.00 -0.01 -0.02 -0.02  0.01 -0.01 -0.04 -0.05 -0.02 -0.02  0.14 -0.02
#> [36,] -0.03  0.04  0.02  0.01  0.03  0.01 -0.01 -0.01  0.00 -0.01 -0.02  0.11
#> [37,] -0.01  0.00  0.02  0.00  0.00 -0.05 -0.02 -0.01 -0.05 -0.03 -0.04  0.04
#> [38,]  0.03  0.02  0.02  0.02  0.01 -0.03  0.01 -0.01 -0.03  0.05 -0.02 -0.01
#> [39,]  0.01 -0.04  0.01 -0.02 -0.03  0.01  0.00  0.01  0.02  0.02  0.02  0.00
#> [40,] -0.02  0.01  0.00 -0.04  0.00  0.03  0.04  0.02 -0.03  0.00  0.04  0.03
#> [41,] -0.03 -0.01  0.01  0.02  0.02 -0.01 -0.04  0.00  0.02  0.01  0.00 -0.02
#> [42,] -0.02  0.03 -0.03 -0.01  0.02  0.01  0.02  0.01 -0.04  0.00  0.02  0.01
#> [43,]  0.00 -0.03 -0.01  0.04  0.00  0.03  0.04 -0.03  0.02  0.04 -0.01 -0.01
#> [44,] -0.06 -0.02  0.03 -0.01  0.01 -0.03 -0.02  0.01  0.03 -0.04  0.01  0.02
#> [45,] -0.02  0.01  0.01 -0.01 -0.01  0.03  0.01  0.00  0.00 -0.01 -0.05 -0.01
#> [46,] -0.01 -0.01  0.02  0.00  0.04  0.01 -0.01 -0.03  0.01 -0.04 -0.03  0.06
#> [47,]  0.03  0.00  0.00  0.06  0.01  0.02  0.03  0.01  0.01  0.00 -0.05  0.01
#> [48,] -0.02 -0.02  0.02  0.03  0.03  0.00  0.06 -0.02 -0.02 -0.02  0.00 -0.01
#> [49,] -0.02 -0.02  0.03  0.01  0.00 -0.01  0.04  0.01  0.02  0.03 -0.03 -0.02
#> [50,] -0.03 -0.03  0.02  0.01 -0.01 -0.03  0.05 -0.02 -0.03  0.00 -0.03 -0.01
#>       [,37] [,38] [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48]
#>  [1,]  0.02 -0.01  0.02  0.03 -0.03  0.04 -0.05 -0.02  0.01  0.05  0.00 -0.05
#>  [2,]  0.02  0.03  0.01  0.04 -0.02  0.02  0.01  0.00 -0.02  0.01 -0.04  0.00
#>  [3,] -0.01  0.00 -0.04 -0.03  0.02  0.05  0.02 -0.01  0.02  0.02  0.01 -0.02
#>  [4,] -0.01  0.01  0.06  0.02  0.02 -0.02  0.00  0.01 -0.03  0.03  0.04  0.00
#>  [5,] -0.01  0.02  0.00  0.00 -0.01 -0.04 -0.01  0.02 -0.02  0.01 -0.04  0.01
#>  [6,]  0.02  0.02  0.01  0.00 -0.03  0.00  0.00  0.03  0.03 -0.02 -0.02 -0.04
#>  [7,]  0.01 -0.03  0.01  0.00 -0.02  0.01 -0.03 -0.02 -0.02  0.01 -0.03 -0.04
#>  [8,] -0.02 -0.01 -0.03  0.01 -0.04 -0.03  0.01  0.02 -0.03  0.00  0.00  0.03
#>  [9,] -0.03  0.00 -0.06 -0.01  0.01  0.02 -0.06 -0.01  0.06 -0.01  0.01  0.01
#> [10,] -0.01  0.05  0.04  0.01 -0.01  0.00  0.00 -0.02 -0.05  0.02  0.01 -0.03
#> [11,] -0.07  0.04  0.01  0.01  0.00  0.00  0.00 -0.01 -0.01 -0.02  0.00 -0.01
#> [12,] -0.02  0.02 -0.05 -0.01 -0.02  0.02  0.01 -0.04 -0.02  0.02  0.01 -0.02
#> [13,]  0.02  0.00  0.03  0.01  0.02 -0.01 -0.04  0.04  0.01 -0.02 -0.01 -0.03
#> [14,] -0.02 -0.04  0.00 -0.02 -0.04  0.01  0.01 -0.01  0.01  0.05  0.01 -0.03
#> [15,]  0.01 -0.01 -0.01 -0.01 -0.02  0.00  0.06 -0.02 -0.02  0.06  0.00  0.02
#> [16,]  0.06 -0.04  0.00 -0.03 -0.07 -0.03  0.02  0.05  0.01  0.04  0.04 -0.02
#> [17,] -0.01 -0.01 -0.04 -0.01 -0.02  0.03  0.05  0.00  0.01 -0.03  0.05  0.04
#> [18,]  0.03  0.07 -0.02  0.05  0.00  0.03  0.00 -0.04 -0.01  0.05  0.01  0.02
#> [19,] -0.01  0.05 -0.03  0.00  0.00  0.05  0.01 -0.04 -0.01  0.02  0.02 -0.02
#> [20,] -0.02 -0.01  0.06  0.03  0.02  0.01  0.00  0.03 -0.04 -0.01  0.03  0.00
#> [21,] -0.01  0.01 -0.04 -0.02 -0.01 -0.02  0.00 -0.03  0.00  0.00 -0.02  0.02
#> [22,] -0.03  0.08 -0.01  0.00  0.01 -0.04  0.00 -0.06 -0.03  0.00  0.01  0.05
#> [23,]  0.02 -0.01 -0.05  0.02 -0.03  0.03  0.00 -0.04  0.00  0.00 -0.04  0.03
#> [24,]  0.05  0.00  0.00  0.00 -0.01 -0.05 -0.04  0.02 -0.01  0.03  0.00  0.03
#> [25,] -0.01  0.03  0.01 -0.02 -0.03 -0.02  0.00 -0.06 -0.02 -0.01  0.03 -0.02
#> [26,]  0.00  0.02 -0.04  0.01 -0.01  0.03 -0.03 -0.02  0.01 -0.01  0.00 -0.02
#> [27,]  0.02  0.02  0.01  0.00  0.01 -0.03 -0.01  0.03  0.01  0.02  0.00  0.02
#> [28,]  0.00  0.02 -0.02 -0.04  0.02 -0.01  0.04 -0.01 -0.01  0.00  0.06  0.03
#> [29,]  0.00  0.01 -0.03  0.00  0.02  0.02  0.00  0.01 -0.01  0.04  0.01  0.03
#> [30,] -0.05 -0.03  0.01  0.03 -0.01  0.01  0.03 -0.03  0.03  0.01  0.02  0.00
#> [31,] -0.02  0.01  0.00  0.04 -0.04  0.02  0.04 -0.02  0.01 -0.01  0.03  0.06
#> [32,] -0.01 -0.01  0.01  0.02  0.00  0.01 -0.03  0.01  0.00 -0.03  0.01 -0.02
#> [33,] -0.05 -0.03  0.02 -0.03  0.02 -0.04  0.02  0.03  0.00  0.01  0.01 -0.02
#> [34,] -0.03  0.05  0.02  0.00  0.01  0.00  0.04 -0.04 -0.01 -0.04  0.00 -0.02
#> [35,] -0.04 -0.02  0.02  0.04  0.00  0.02 -0.01  0.01 -0.05 -0.03 -0.05  0.00
#> [36,]  0.04 -0.01  0.00  0.03 -0.02  0.01 -0.01  0.02 -0.01  0.06  0.01 -0.01
#> [37,]  0.14  0.01 -0.01  0.00 -0.05  0.00 -0.05  0.01  0.02  0.01  0.00 -0.01
#> [38,]  0.01  0.19 -0.01  0.00  0.00 -0.01 -0.02 -0.12 -0.03  0.03  0.01 -0.04
#> [39,] -0.01 -0.01  0.12  0.03  0.01 -0.04  0.01  0.03 -0.03 -0.01  0.00 -0.04
#> [40,]  0.00  0.00  0.03  0.13 -0.06  0.05 -0.02 -0.01 -0.02  0.00 -0.04 -0.01
#> [41,] -0.05  0.00  0.01 -0.06  0.17 -0.04 -0.02  0.03  0.01 -0.02  0.00  0.02
#> [42,]  0.00 -0.01 -0.04  0.05 -0.04  0.12  0.02 -0.04  0.01  0.01 -0.01 -0.02
#> [43,] -0.05 -0.02  0.01 -0.02 -0.02  0.02  0.17 -0.01 -0.04  0.02  0.04  0.02
#> [44,]  0.01 -0.12  0.03 -0.01  0.03 -0.04 -0.01  0.17  0.03 -0.07 -0.01  0.03
#> [45,]  0.02 -0.03 -0.03 -0.02  0.01  0.01 -0.04  0.03  0.09 -0.02 -0.01 -0.01
#> [46,]  0.01  0.03 -0.01  0.00 -0.02  0.01  0.02 -0.07 -0.02  0.19  0.03 -0.02
#> [47,]  0.00  0.01  0.00 -0.04  0.00 -0.01  0.04 -0.01 -0.01  0.03  0.10  0.02
#> [48,] -0.01 -0.04 -0.04 -0.01  0.02 -0.02  0.02  0.03 -0.01 -0.02  0.02  0.12
#> [49,] -0.05 -0.01  0.00  0.03 -0.01  0.02  0.06  0.03  0.01  0.01  0.01  0.00
#> [50,]  0.00  0.01  0.02  0.01  0.00 -0.04 -0.03  0.00  0.02  0.05 -0.01  0.03
#>       [,49] [,50]
#>  [1,] -0.02  0.00
#>  [2,] -0.05  0.03
#>  [3,] -0.01 -0.03
#>  [4,] -0.02  0.05
#>  [5,]  0.00  0.01
#>  [6,]  0.02 -0.01
#>  [7,] -0.03  0.02
#>  [8,]  0.00 -0.01
#>  [9,] -0.03  0.01
#> [10,] -0.02 -0.02
#> [11,]  0.02  0.03
#> [12,]  0.04 -0.04
#> [13,]  0.02 -0.01
#> [14,]  0.02  0.01
#> [15,] -0.03  0.01
#> [16,]  0.01  0.01
#> [17,]  0.04 -0.02
#> [18,]  0.06  0.03
#> [19,]  0.05  0.00
#> [20,]  0.01  0.00
#> [21,] -0.03  0.04
#> [22,]  0.00 -0.02
#> [23,] -0.02 -0.01
#> [24,] -0.01  0.03
#> [25,] -0.02 -0.03
#> [26,] -0.02 -0.03
#> [27,]  0.03  0.02
#> [28,]  0.01  0.01
#> [29,]  0.00 -0.01
#> [30,] -0.01 -0.03
#> [31,]  0.04  0.05
#> [32,]  0.01 -0.02
#> [33,]  0.02 -0.03
#> [34,]  0.03  0.00
#> [35,] -0.03 -0.03
#> [36,] -0.02 -0.01
#> [37,] -0.05  0.00
#> [38,] -0.01  0.01
#> [39,]  0.00  0.02
#> [40,]  0.03  0.01
#> [41,] -0.01  0.00
#> [42,]  0.02 -0.04
#> [43,]  0.06 -0.03
#> [44,]  0.03  0.00
#> [45,]  0.01  0.02
#> [46,]  0.01  0.05
#> [47,]  0.01 -0.01
#> [48,]  0.00  0.03
#> [49,]  0.16  0.00
#> [50,]  0.00  0.16