Perform SiMLR Analysis on Multimodal ANTsPyMM Data

This function processes multimodal data using SiMLR. It is designed to be flexible, allowing for various preprocessing steps and analysis options. The analysis can be adjusted through multiple parameters, offering control over the inclusion of certain data types, permutation testing, and more.

antspymm_simlr(
  blaster,
  select_training_boolean,
  connect_cog,
  energy = c("cca", "reg", "lrr", "regression", "base.rand", "base.pca"),
  nsimlr,
  constraint,
  covariates = "1",
  myseed = 3,
  doAsym = TRUE,
  returnidps = FALSE,
  restrictDFN = FALSE,
  resnetGradeThresh = 1.02,
  doperm = FALSE,
  exclusions = NULL,
  inclusions = NULL,
  sparseness = NULL,
  iterations = NULL,
  path_modeling = NULL,
  sparsenessAlg = NA,
  verbose = FALSE
)

Arguments

blaster

A dataframe containing multimodal data for analysis.

select_training_boolean

boolean vector to define which entries are in training data

connect_cog

Vector of column names to be treated as a special target matrix; often used for cognitive data and in a superivsed variant of simlr. Exclude this argument if this is unclear.

energy

The type of energy model to use for similarity analysis. Defaults to 'reg'.

nsimlr

Number of components.

constraint

orthogonality constraint of the form constraintxFloatWeightEnergyxFloatWeightGrad where constraints is ortho, Stiefel or Grassmann or GrassmannInv

covariates

any covariates to adjust training matrices. if covariates is set to 'mean' then the rowwise mean will be factored out of each matrix. this can be a vector e.g. c('center','scale','rank'). pass the name opt to antspymm_simlr_update_residuals to have the function print the options.

myseed

Seed for random number generation to ensure reproducibility. Defaults to 3.

doAsym

integer 0 for FALSE, 1 for TRUE and 2 for separate matrices for asymm variables.

returnidps

Logical indicating whether to return the intermediate processing steps' results. Defaults to FALSE.

restrictDFN

Logical indicating whether to restrict analysis to default network features. Defaults to FALSE.

resnetGradeThresh

image quality threshold (higher better).

doperm

Logical indicating whether to perform permutation tests. Defaults to FALSE. Will randomize image features in the training data and thus leads to "randomized" but still regularized projections.

exclusions

vector of strings to exclude from predictors

inclusions

vector of strings to include in predictors

sparseness

vector or scalar value to set sparseness

iterations

int value to set max iterations

path_modeling

the result of a call to simlr_path_models(n)

sparsenessAlg

NA is default otherwise basic, spmp or orthorank

verbose

boolean

Value

A list containing the results of the similarity analysis and related data.

Examples

# Example usage:
# result <- antspymm_simlr(dataframe)