Euler Character

Mostly an for internal use in conjunction with resels. Calculates the euler characteristic at a threshold level.

euler(u, df, fieldType)

Arguments

u
statistical value (typically the maxima of a cluster or statistical field)

df
degrees of freedom expressed as df = c(degrees of interest, degrees of error)

fieldType

T: T-field

F: F-field

X: Chi-square field'

Z: Gaussian field

Value

A vector of estimated euler characteristics for dimensions 0:D

References

Worlsey K.J., (1996) A Unified Statistical Approach for Determining Significant Signals in Images of Cerebral Activation.

See also

rftPval, resels

Examples

mask <- getMask(  antsImageRead( getANTsRData( 'r16' )  ) )
myresels <- resels(mask, c(1, 1))
# pretend these results are from a t-statistical field fitted with 10 df to
# find a peak value.
df <- c(1, 10) # made up degrees of freedom
ec <- euler(4, df, fieldType = "T") # peak value is 4
ez <- euler(3, df, fieldType = "T") # we thresholded at 3
# uncorrected peak-level p-value
pvox <- sum(ec * myresels)/sum(ez * myresels)