Label multivariate components by an anatomical coordinate system.

This function will provide a mapping that labels the list of input images and each of their blobs.

getMultivariateTemplateCoordinates(imageSetToBeLabeledIn,
  templateWithLabels, labelnames = NA, outprefix = NA,
  convertToTal = FALSE, pvals = NA, threshparam = 1,
  clustparam = 250, identifier)

Arguments

imageSetToBeLabeledIn
a template paired with (most likely) the output of a multivariate sparse decomposition or (alternatively) could be just a statistical map with zeroes in non-interesting areas

templateWithLabels
e..g the mni image and brodmann label set

labelnames
a list of names for the labels

outprefix
if output to a file, provide file prefix

convertToTal
bool, return talairach coordinates

pvals
the already computed pvalue for each component

threshparam
for pvals

clustparam
for clusters

identifier
unique ID for this study

Value

The output point coordinates are in approximate Talairach / MNI (or whatever) template space.

Details

Uses getTemplateCoordinates as a sub-routine.

TBN

Examples

# NOT RUN {
 tem<-antsImageRead( getANTsRData("ch2") )
 temlab<-antsImageRead( getANTsRData("ch2b")  )
 temlab2<-antsImageRead( getANTsRData("ch2a")  )
 # try getANTsRData if you have www access
 mymni<-list( antsImageRead(getANTsRData("mni"),3),
       antsImageRead(getANTsRData("mnib"),3),
       antsImageRead(getANTsRData("mnia"),3) )
 mytem<-list( smoothImage(tem,3) ,temlab,temlab2)
 # mynetworkdescriptor<-getMultivariateTemplateCoordinates(
 #  mytem, mymni , convertToTal = TRUE , pvals=c(0.01,0.05) )
# }