Compute cortical thickness using the DiReCT algorithm.

Diffeomorphic registration-based cortical thickness based on probabilistic segmentation of an image. This is an optimization algorithm.

kellyKapowski(
  s,
  g,
  w,
  its = 50,
  r = 0.025,
  m = 1.5,
  x = TRUE,
  e = FALSE,
  t = NULL,
  timeSigma = 1,
  verbose = FALSE,
  ...
)

Arguments

s
segmentation image

g
gray matter probability image

w
white matter probability image

its
convergence params - controls iterations

r
gradient descent update parameter

m
gradient field smoothing parameter

x
matrix-based smoothing

e
restrict deformation boolean

t
time spacing, a vector equal to the number of time dimensions

timeSigma,
a scalar sigma value for distances between time points

verbose
boolean

...
anything else, see KK help in ANTs

Value

thickness antsImage

Examples

img<-antsImageRead( getANTsRData("r16") ,2)
img<-resampleImage(img,c(64,64),1,0)
mask<-getMask( img )
segs<-kmeansSegmentation( img, k=3, kmask = mask)
thk<-kellyKapowski( s=segs$segmentation, g=segs$probabilityimages[[2]],
  w=segs$probabilityimages[[3]],its=45,r=0.5,m=1 )