powers_areal_mni_itk

A data frame providing coordinates for the nodes identified by Powers at al. Coordinates have been modified from nifti space to ITK space

powers_areal_mni_itk

Format

A data frame listing the following variables.

x

x coordinate

y

y coordinate

z

z coordinate

ROI

unique integer for each node

SystemLabel

unique integer for each system

SystemName

unique name for each system

Color

name of color for system

r

red value for system

g

green value for system

b

blue value for system

Anatomy

anatomical location as determined by OASIS labeling

Lobe

lobe location of node

Brodmann

Broadmann area number

AAL

AAL region label

#' @references http://www.nil.wustl.edu/labs/petersen/Resources.html

Examples

data(powers_areal_mni_itk)
if (FALSE) {
  fixed <- antsImageRead( getANTsRData("ch2") )
  moving <- antsImageRead( getANTsRData("mni") )
  mytx <- antsRegistration( fixed=fixed, moving=moving,
  typeofTransform = c("SyN") )
  data( "powers_areal_mni_itk", package = "ANTsR", envir = environment() )
  coords = powers_areal_mni_itk[,1:3]
  ch2reg   = antsRegistration( fixed, moving, typeofTransform = "SyN" )
  coordsw <- antsApplyTransformsToPoints(
  dim=3, points=coords,
  transformlist=ch2reg$fwdtransforms,
  whichtoinvert=c(FALSE,FALSE) )
  ptrd = 3
  powersLabels = makePointsImage( coordsw, moving, radius = pard )
  plot( moving, powersLabels, axis=3, nslices=30 )
}