Compute the low-dimensional u matrix for simlr

simlr minimizes reconstruction error across related modalities. One crucial component of the reconstruction is the low-dimensional cross-modality basis. This function computes that basis, given a mixing algorithm.

simlrU(
  projections,
  mixingAlgorithm,
  initialW,
  orthogonalize = FALSE,
  connectors = NULL
)

Arguments

projections
A list that contains the low-dimensional projections.

mixingAlgorithm
the elected mixing algorithm. see simlr. can be 'svd', 'ica', 'rrpca-l', 'rrpca-s', 'pca', 'stochastic' or 'avg'.

initialW
initialization matrix size n by k for fastICA.

orthogonalize
boolean

connectors
a list ( length of projections or number of modalities ) that indicates which modalities should be paired with current modality

Value

u matrix for modality i

See also

Examples

set.seed(1500)
nsub = 25
npix = c(100,200,133)
nk = 5
outcome = matrix(rnorm( nsub * nk ),ncol=nk)
outcome1 = matrix(rnorm( nsub * nk ),ncol=nk)
outcome2 = matrix(rnorm( nsub * nk ),ncol=nk)
u = simlrU( list( outcome, outcome1, outcome2 ), 2, 'avg' )