Symmetric multivariate, penalized image-based linear regression model (symilr) for N modalities

This function simplifies calculating image-wide multivariate beta maps from that is similar to CCA.

symilr(voxmats, smoothingMatrices, iterations = 10, gamma = 1e-06,
  sparsenessQuantiles, positivities, initialUMatrix, orthogonalize = TRUE,
  repeatedMeasures = NA, verbose = FALSE)

Arguments

voxmats
A list that contains the named x and y matrices.

smoothingMatrices
list of matrices, allows parameter smoothing, should be square and same size as input matrices

iterations
number of gradient descent iterations

gamma
step size for gradient descent

sparsenessQuantiles
vector of quantiles to control sparseness - higher is sparser

positivities
vector that sets for each matrix if we restrict to positive or negative solution (beta) weights. choices are positive, negative or either as expressed as a string.

initialUMatrix
list of initialization matrix size n by k for each modality. If this is set to a scalar, or is missing, a random matrix will be used.

orthogonalize
boolean to control whether we orthogonalize the solutions explicitly

repeatedMeasures
list of repeated measurement identifiers. this will allow estimates of per identifier intercept.

verbose
boolean to control verbosity of output

Value

A list of u, x, y, z etc related matrices.

See also

Examples

set.seed(1500)
nsub = 12
npix = 100
outcome = rnorm( nsub )
covar = rnorm( nsub )
mat = replicate( npix, rnorm( nsub ) )
mat2 = replicate( npix + 10, rnorm( nsub ) )
mat3 = replicate( npix + 10, rnorm( nsub ) )
nk = 3
result = symilr(
  list( vox = mat, vox2 = mat2, vox3 = mat3 ), initialUMatrix = nk  )

#> length( initialUMatrix ) != length( voxmats ) -- initializing with random matrix